refactor(tentacle): extract dispatch and review seams
