+++ CONTRIBUTING.md
@@ -151,6 +151,7 @@ sync-config.py  sync-daemon.py  sync-status.py
 migrate.py  generate-summary.py
 briefing.py  learn.py  query-session.py  extract-knowledge.py
 build-session-index.py  tentacle.py  _tentacle_core.py  _tentacle_goal.py  _tentacle_pr.py
+_tentacle_dispatch.py  _tentacle_review.py
 checkpoint-diff.py  checkpoint-restore.py  checkpoint-save.py
 tests/test_browse_search_v2.py
 browse/  hooks/  scripts/
