+++ hooks/pre-commit
@@ -126,6 +126,8 @@ def in_python_cleanliness_surface(path: str) -> bool:
         "_tentacle_core.py",
         "_tentacle_goal.py",
         "_tentacle_pr.py",
+        "_tentacle_dispatch.py",
+        "_tentacle_review.py",
         "checkpoint-diff.py",
         "checkpoint-restore.py",
         "checkpoint-save.py",
