+++ tentacle.py
@@ -72,8 +72,10 @@ from contextlib import contextmanager
 from datetime import datetime, timezone
 from pathlib import Path
 
+import _tentacle_dispatch as _dispatch
 import _tentacle_goal as _goal
 import _tentacle_pr as _pr
+import _tentacle_review as _review
 from _tentacle_core import (
     _DISPATCHED_MARKER_NAME,
     _DISPATCHED_MARKER_PATH,
@@ -98,6 +100,53 @@ from _tentacle_core import (
     get_tentacles_dir,
     parse_todos,
     render_todos,
+)
+from _tentacle_dispatch import (
+    FULL_CONTEXT_DISPATCH_MODE_NAME,
+    POINTER_DISPATCH_MODE_NAME,
+    POINTER_PROMPT_REDUCTION_TARGET_PERCENT,
+    _agent_profile_candidates,
+    _agent_profile_meta,
+    _apply_context_packet_template,
+    _blocker_context_from_meta,
+    _build_context_packet,
+    _build_runtime_bundle,
+    _bundle_enabled,
+    _context_excerpt,
+    _dispatch_context_mode,
+    _dispatch_prompt_size_stats,
+    _extract_pack_entries,
+    _fetch_recall_pack_json,
+    _load_agent_profile,
+    _load_context_packet_template,
+    _load_latest_checkpoint_context,
+    _normalize_agent_profile_id,
+    _pack_payload_has_signal,
+    _parse_agent_frontmatter,
+    _parse_frontmatter_scalar,
+    _profile_list,
+    _project_conventions_summary,
+    _recall_pack_summary,
+    _render_agent_profile_section,
+    _render_checkpoint_context,
+    _render_dispatch_context,
+    _render_dispatch_live_briefing_section,
+    _render_dispatch_mode_section,
+    _render_dispatch_prompt_size_section,
+    _render_knowledge_evidence,
+    _render_profile_list,
+    _render_recall_payload,
+    _render_swarm_prompt,
+    _resolve_agent_profile_from_meta,
+    _run_briefing,
+    _run_briefing_for_task,
+    _scope_items,
+    _scope_summary,
+    _upsert_auto_recall_block,
+    cmd_bundle,
+    cmd_next_step,
+    cmd_resume,
+    cmd_swarm,
 )
 from _tentacle_goal import (
     _GOAL_LOCK_POLL_S,
@@ -195,6 +244,50 @@ from _tentacle_pr import (
     _pr_run_subprocess_safe,
     cmd_pr,
 )
+from _tentacle_review import (
+    REVIEW_LOOP_ACTIONABLE_CLASSIFICATIONS,
+    REVIEW_LOOP_BUILD_ERROR_COMMAND_TOKENS,
+    REVIEW_LOOP_BUILD_ERROR_PATTERNS,
+    REVIEW_LOOP_CLASSIFICATIONS,
+    REVIEW_LOOP_TEST_FAILURE_PATTERNS,
+    REVIEWER_BUNDLE_DIRNAME,
+    REVIEWER_FINDINGS_FILENAME,
+    REVIEWER_PENDING_VALUES,
+    REVIEWER_SAFE_FALSE,
+    REVIEWER_SAFE_TRUE,
+    _build_reviewer_bundle,
+    _extract_markdown_section,
+    _render_dispatch_reviewer_prompt,
+    _require_done_handoff,
+    _review_loop_actionable_count,
+    _review_loop_append_history,
+    _review_loop_changed_files,
+    _review_loop_classify_failure,
+    _review_loop_collect_baseline_failures,
+    _review_loop_create_resolver_tentacle,
+    _review_loop_dispatch_resolver,
+    _review_loop_failure_hash,
+    _review_loop_handle_reviewer_findings,
+    _review_loop_history,
+    _review_loop_is_preexisting_failure,
+    _review_loop_is_test_file,
+    _review_loop_latest_verify_command,
+    _review_loop_next_resolver_name,
+    _review_loop_output_excerpt,
+    _review_loop_read_output,
+    _reviewer_bundle_dir,
+    _reviewer_collect_diff,
+    _reviewer_collect_untracked_patch,
+    _reviewer_findings_path,
+    _reviewer_findings_template,
+    _reviewer_load_findings,
+    _reviewer_parse_findings,
+    _reviewer_render_spec_text,
+    _reviewer_step_paths,
+    _reviewer_worktree_path,
+    cmd_dispatch_reviewer,
+    cmd_review_loop,
+)
 
 # Fix Windows console encoding
 if os.name == "nt":
@@ -219,6 +312,7 @@ if os.name == "nt":
 # | handoff-complete | handoff/complete/audit tests | handoff parsing, completion, metrics helpers |
 # | stop-event-cleanup | Rust hook runner fallback and session lifecycle hooks | marker-cleanup CLI behavior |
 # | pr-automation | tests/test_tentacle_pr.py | _pr_* helpers and cmd_pr |
+# | review-loop | tests/test_tentacle_runtime.py | reviewer bundle helpers, _review_loop_* helpers, cmd_dispatch_reviewer, cmd_review_loop |
 
 # ---------------------------------------------------------------------------
 # Structured handoff contract constants
@@ -239,33 +333,6 @@ HANDOFF_RESETTABLE_STATUSES: frozenset[str] = frozenset({"BLOCKED", "AMBIGUOUS",
 SCOPE_ESCALATION_FILE_THRESHOLD = 4
 SCOPE_REDUCTION_FILE_THRESHOLD = 2
 SCOPE_ESCALATION_SPLIT_CHILDREN = 2
-REVIEW_LOOP_CLASSIFICATIONS: frozenset[str] = frozenset(
-    {"PRE_EXISTING", "FLAKY", "BUILD_ERROR", "NEW_TEST_WRONG", "REGRESSION"}
-)
-REVIEW_LOOP_ACTIONABLE_CLASSIFICATIONS: frozenset[str] = frozenset({"BUILD_ERROR", "NEW_TEST_WRONG", "REGRESSION"})
-REVIEW_LOOP_BUILD_ERROR_COMMAND_TOKENS: tuple[str, ...] = ("py_compile", "ruff", "mypy", "compileall")
-REVIEW_LOOP_BUILD_ERROR_PATTERNS: tuple[str, ...] = (
-    r"\bSyntaxError\b",
-    r"\bIndentationError\b",
-    r"\bImportError\b",
-    r"\bModuleNotFoundError\b",
-    r"No module named",
-)
-REVIEW_LOOP_TEST_FAILURE_PATTERNS: tuple[str, ...] = (
-    r"\bAssertionError\b",
-    r"\bFAILED\b",
-    r"\bassert\b",
-    r"\bexpected\b",
-)
-REVIEWER_BUNDLE_DIRNAME = "reviewer-bundle"
-REVIEWER_FINDINGS_FILENAME = "reviewer-findings.md"
-REVIEWER_SAFE_TRUE: frozenset[str] = frozenset({"YES", "TRUE"})
-REVIEWER_SAFE_FALSE: frozenset[str] = frozenset({"NO", "FALSE"})
-REVIEWER_PENDING_VALUES: frozenset[str] = frozenset({"PENDING", "UNKNOWN", "TBD"})
-POINTER_DISPATCH_MODE_NAME = "pointer_bundle"
-FULL_CONTEXT_DISPATCH_MODE_NAME = "full_context_inline"
-POINTER_PROMPT_REDUCTION_TARGET_PERCENT = 30.0
-
 # ---------------------------------------------------------------------------
 # Automation constants
 # ---------------------------------------------------------------------------
@@ -284,756 +351,7 @@ AUTOMATION_EVENT_MAP = {
 # --- Commands ---
 
 
-def _run_briefing(query: str) -> str:
-    """Run briefing.py with a text query and return compact output. Returns empty string on failure."""
-    if not BRIEFING_PY.exists():
-        return ""
-    try:
-        result = subprocess.run(
-            [sys.executable, str(BRIEFING_PY), query, "--compact", "--limit", "3"],
-            capture_output=True,
-            text=True,
-            encoding="utf-8",
-            errors="replace",  # P1-3: prevent UnicodeDecodeError on Windows cp1252
-            timeout=15,
-        )
-        output = result.stdout.strip()
-        if output and "No relevant" not in output and len(output) > 20:
-            return output
-    except (subprocess.TimeoutExpired, Exception):
-        pass
-    return ""
-
-
-def _render_knowledge_evidence(
-    entries: list[dict],
-    *,
-    task_id: str = "",
-    file_matches: list[dict] | None = None,
-) -> str:
-    """Render deterministic compact evidence block for prompt injection."""
-
-    def _source_label(entry: dict) -> str:
-        src = entry.get("source_document") or {}
-        if not isinstance(src, dict):
-            return ""
-        doc_type = str(src.get("doc_type") or "").strip()
-        if not doc_type:
-            return ""
-        section = str(src.get("section") or "").strip()
-        seq = src.get("seq")
-        file_path = str(src.get("file_path") or "").strip()
-        title = str(src.get("title") or "").strip()
-        if doc_type == "checkpoint" and seq:
-            label = f"checkpoint #{seq}"
-        elif file_path:
-            label = f"{doc_type} / {Path(file_path).name}"
-        elif title:
-            label = f"{doc_type} / {title[:80]}"
-        else:
-            label = doc_type
-        if section:
-            label = f"{label} / {section}"
-        return label[:120]
-
-    refs = entries[:5]
-    if not refs:
-        return ""
-    lines = ["[KNOWLEDGE EVIDENCE]"]
-    if task_id:
-        lines.append(f"Task: {task_id}")
-    for e in refs:
-        eid = e.get("id", "?")
-        cat = e.get("category", "unknown")
-        title = e.get("title", "(no title)")
-        lines.append(f"- #{eid} [{cat}] {title}")
-    labels: list[str] = []
-    for e in refs:
-        label = _source_label(e)
-        if not label or label in labels:
-            continue
-        labels.append(label)
-        if len(labels) >= 2:
-            break
-    if labels:
-        lines.append(f"From: {'; '.join(labels)}")
-    file_paths: list[str] = []
-    for fm in file_matches or []:
-        path = str(fm.get("file_or_module", "")).strip()
-        if path and path not in file_paths:
-            file_paths.append(path)
-        if len(file_paths) >= 3:
-            break
-    if file_paths:
-        lines.append(f"Files: {', '.join(file_paths)}")
-    first_entry = refs[0]
-    first_id = first_entry.get("id", "?")
-    drilldowns = [f"query-session.py --detail {first_id}"]
-    first_related = first_entry.get("related_entry_ids")
-    if isinstance(first_related, list) and len(first_related) > 0:
-        drilldowns.append(f"query-session.py --related {first_id}")
-    if task_id:
-        drilldowns.append(f"query-session.py --task {task_id!r}")
-    if file_paths:
-        drilldowns.append(f"query-session.py {file_paths[0]!r}")
-    lines.append(f"Drilldown: {' | '.join(drilldowns)}")
-    return "\n".join(lines)
-
-
-def _extract_pack_entries(pack_data: dict) -> list[dict]:
-    """Extract ordered reference entries from briefing --pack payload."""
-    entries = pack_data.get("entries", {})
-    out = []
-    for category in ("mistake", "pattern", "decision", "tool"):
-        for entry in entries.get(category, []):
-            out.append(
-                {
-                    "id": entry.get("id", "?"),
-                    "category": entry.get("category", category),
-                    "title": entry.get("title", "(no title)"),
-                    "source_document": entry.get("source_document"),
-                    "related_entry_ids": entry.get("related_entry_ids", []),
-                }
-            )
-    return out
-
-
-def _run_briefing_for_task(task_id: str, fallback_query: str = "") -> str:
-    """Load evidence block for task recall using task-json then pack fallback."""
-    recall_pack_data, recall_source_mode = _fetch_recall_pack_json(task_id, fallback_query=fallback_query)
-    return _render_recall_payload(task_id, recall_pack_data, recall_source_mode)
-
-
-def _pack_payload_has_signal(pack_data: dict) -> bool:
-    """Return True when a --pack payload carries actionable recall content."""
-    entries = pack_data.get("entries", {})
-    if any(entries.get(cat) for cat in ("mistake", "pattern", "decision", "tool")):
-        return True
-    for key in ("task_matches", "file_matches", "past_work", "risk"):
-        if pack_data.get(key):
-            return True
-    return bool(pack_data.get("next_open"))
-
-
-def _render_recall_payload(task_id: str, recall_data: dict, source_mode: str | None) -> str:
-    """Render a fetched recall payload into the bounded prose evidence block."""
-    if not recall_data or not source_mode:
-        return ""
-    if source_mode == "task_json":
-        tagged = recall_data.get("tagged_entries", [])
-        related = recall_data.get("related_entries", [])
-        if not (tagged or related):
-            return ""
-        task_entries = [
-            {
-                "id": e.get("id", "?"),
-                "category": e.get("category", "unknown"),
-                "title": e.get("title", "(no title)"),
-                "source_document": e.get("source_document"),
-                "related_entry_ids": e.get("related_entry_ids", []),
-            }
-            for e in [*tagged, *related]
-        ]
-        return _render_knowledge_evidence(task_entries, task_id=task_id)
-    if source_mode == "pack":
-        pack_entries = _extract_pack_entries(recall_data)
-        return _render_knowledge_evidence(
-            pack_entries,
-            file_matches=recall_data.get("file_matches", []),
-        )
-    return ""
-
-
-def _fetch_recall_pack_json(task_id: str, fallback_query: str = "") -> tuple[dict, str | None]:
-    """Fetch machine-readable recall JSON for task_id from briefing.py.
-
-    Tries --task --json first (source_mode="task_json"), then --pack fallback
-    (source_mode="pack").  Returns ({}, None) when both sources are empty or
-    briefing.py is unavailable.
-    """
-    if not BRIEFING_PY.exists():
-        return {}, None
-    # Try task-json first
-    try:
-        result = subprocess.run(
-            [sys.executable, str(BRIEFING_PY), "--task", task_id, "--json"],
-            capture_output=True,
-            text=True,
-            encoding="utf-8",
-            errors="replace",
-            timeout=15,
-        )
-        if result.returncode == 0 and result.stdout.strip():
-            data = json.loads(result.stdout)
-            if data.get("tagged_entries") or data.get("related_entries"):
-                return data, "task_json"
-    except (subprocess.TimeoutExpired, json.JSONDecodeError, OSError):
-        pass
-    # Fallback to --pack
-    if fallback_query:
-        try:
-            result = subprocess.run(
-                [
-                    sys.executable,
-                    str(BRIEFING_PY),
-                    fallback_query,
-                    "--pack",
-                    "--limit",
-                    "3",
-                ],
-                capture_output=True,
-                text=True,
-                encoding="utf-8",
-                errors="replace",
-                timeout=15,
-            )
-            if result.returncode == 0 and result.stdout.strip():
-                data = json.loads(result.stdout)
-                if _pack_payload_has_signal(data):
-                    return data, "pack"
-        except (subprocess.TimeoutExpired, json.JSONDecodeError, OSError):
-            pass
-    return {}, None
-
-
-def _upsert_auto_recall_block(context_text: str, recall_content: str) -> str:
-    """Insert/replace a single auto-managed recall block in CONTEXT.md."""
-    block = f"{AUTO_RECALL_START}\n{recall_content}\n{AUTO_RECALL_END}"
-    pattern = re.compile(
-        rf"{re.escape(AUTO_RECALL_START)}.*?{re.escape(AUTO_RECALL_END)}",
-        flags=re.DOTALL,
-    )
-    if pattern.search(context_text):
-        return pattern.sub(block, context_text, count=1)
-    prefix = "" if not context_text or context_text.endswith("\n") else "\n"
-    return f"{context_text}{prefix}\n{block}\n"
-
-
-def _render_checkpoint_context(data: dict) -> str:
-    """Render a concise checkpoint context block from checkpoint JSON.
-
-    Sources only real fields: seq, title, and a small subset of useful sections.
-    """
-    seq = data.get("seq", "?")
-    title = data.get("title", "unknown")
-    sections = data.get("sections", {})
-    lines = [f"### Latest Checkpoint (#{seq}: {title})", ""]
-    for key in ("overview", "work_done", "next_steps"):
-        text = sections.get(key, "").strip()
-        if text:
-            snippet = text[:300] + ("…" if len(text) > 300 else "")
-            label = key.replace("_", " ").title()
-            lines.append(f"**{label}:** {snippet}")
-            lines.append("")
-    return "\n".join(lines).strip()
-
-
-def _load_latest_checkpoint_context() -> str:
-    """Load latest checkpoint and render a concise context block.
-
-    Returns empty string if no checkpoint exists or on any error.
-    """
-    if not CHECKPOINT_RESTORE_PY.exists():
-        return ""
-    try:
-        result = subprocess.run(
-            [
-                sys.executable,
-                str(CHECKPOINT_RESTORE_PY),
-                "--export",
-                "latest",
-                "--format",
-                "json",
-            ],
-            capture_output=True,
-            text=True,
-            encoding="utf-8",
-            errors="replace",  # P1-3
-            timeout=15,
-        )
-        if result.returncode != 0 or not result.stdout.strip():
-            return ""
-        data = json.loads(result.stdout)
-        return _render_checkpoint_context(data)
-    except (subprocess.TimeoutExpired, json.JSONDecodeError, Exception):
-        return ""
-
-
-def _bundle_enabled(args) -> bool:
-    """Return whether dispatch should materialize a runtime bundle.
-
-    The CLI parser defaults this to True for swarm/dispatch. Older direct
-    callers/tests that do not provide the attribute retain the historical
-    behavior.
-    """
-    return bool(getattr(args, "bundle", False))
-
-
-def _normalize_agent_profile_id(profile_id: str) -> str:
-    """Return a safe profile slug for .agent.md lookup."""
-    slug = str(profile_id or "").strip()
-    if slug.endswith(".agent.md"):
-        slug = slug[: -len(".agent.md")]
-    elif slug.endswith(".md"):
-        slug = slug[: -len(".md")]
-    if not re.fullmatch(r"[A-Za-z0-9][A-Za-z0-9_-]*", slug or ""):
-        raise ValueError("Agent profile id must be a slug containing only letters, numbers, underscores, and hyphens.")
-    return slug
-
-
-def _parse_frontmatter_scalar(value: str):
-    """Parse the scalar subset used by .agent.md frontmatter."""
-    raw = value.strip()
-    if raw == "":
-        return ""
-    if raw.startswith("[") and raw.endswith("]"):
-        try:
-            parsed = ast.literal_eval(raw)
-        except (SyntaxError, ValueError):
-            return raw
-        if isinstance(parsed, list):
-            return [str(item) for item in parsed]
-        return raw
-    if (raw.startswith("'") and raw.endswith("'")) or (raw.startswith('"') and raw.endswith('"')):
-        return raw[1:-1]
-    lower = raw.lower()
-    if lower == "true":
-        return True
-    if lower == "false":
-        return False
-    return raw
-
-
-def _parse_agent_frontmatter(path: Path) -> dict:
-    """Parse the YAML-frontmatter subset used by bundled agent profiles."""
-    text = path.read_text(encoding="utf-8", errors="replace")
-    lines = text.splitlines()
-    if not lines or lines[0].strip() != "---":
-        return {}
-    end_index = None
-    for idx in range(1, len(lines)):
-        if lines[idx].strip() == "---":
-            end_index = idx
-            break
-    if end_index is None:
-        return {}
-
-    data: dict = {}
-    frontmatter = lines[1:end_index]
-    i = 0
-    while i < len(frontmatter):
-        line = frontmatter[i]
-        stripped = line.strip()
-        if not stripped or stripped.startswith("#") or line.startswith((" ", "\t")):
-            i += 1
-            continue
-        if ":" not in line:
-            i += 1
-            continue
-
-        key, raw_value = line.split(":", 1)
-        key = key.strip()
-        value = raw_value.strip()
-        if not key:
-            i += 1
-            continue
-
-        if value in {"|", ">"}:
-            block_lines: list[str] = []
-            i += 1
-            while i < len(frontmatter):
-                block_line = frontmatter[i]
-                if block_line.strip() and not block_line.startswith((" ", "\t")):
-                    break
-                block_lines.append(block_line[2:] if block_line.startswith("  ") else block_line.lstrip())
-                i += 1
-            data[key] = "\n".join(block_lines).rstrip()
-            continue
-
-        if value == "":
-            items: list = []
-            i += 1
-            while i < len(frontmatter):
-                item_line = frontmatter[i]
-                if item_line.strip() and not item_line.startswith((" ", "\t")):
-                    break
-                item_stripped = item_line.strip()
-                if item_stripped.startswith("- "):
-                    items.append(_parse_frontmatter_scalar(item_stripped[2:]))
-                i += 1
-            data[key] = items if items else ""
-            continue
-
-        data[key] = _parse_frontmatter_scalar(value)
-        i += 1
-
-    return data
-
-
-def _agent_profile_candidates(profile_id: str, git_root: Path | None = None) -> list[Path]:
-    """Return profile lookup candidates in project-first order."""
-    slug = _normalize_agent_profile_id(profile_id)
-    candidates: list[Path] = []
-    if git_root:
-        candidates.extend(
-            [
-                git_root / ".github" / "agents" / f"{slug}.agent.md",
-                git_root / ".github" / "agents" / f"{slug}.md",
-            ]
-        )
-    candidates.extend(
-        [
-            AGENT_PROFILE_REFERENCE_DIR / f"{slug}.agent.md",
-            AGENT_PROFILE_REFERENCE_DIR / f"{slug}.md",
-        ]
-    )
-    return candidates
-
-
-def _load_agent_profile(profile_id: str, git_root: Path | None = None) -> dict:
-    """Load a project or bundled agent profile from .agent.md frontmatter."""
-    slug = _normalize_agent_profile_id(profile_id)
-    candidates = _agent_profile_candidates(slug, git_root)
-    for candidate in candidates:
-        if not candidate.is_file():
-            continue
-        profile = _parse_agent_frontmatter(candidate)
-        if not profile:
-            raise ValueError(f"Agent profile file has no YAML frontmatter: {candidate}")
-        profile["profile_id"] = str(profile.get("profile_id") or slug)
-        profile["role"] = str(profile.get("role") or profile.get("name") or profile["profile_id"])
-        profile["agent_type"] = str(profile.get("agent_type") or profile["profile_id"])
-        profile["profile_path"] = str(candidate)
-        return profile
-
-    searched = ", ".join(str(path) for path in candidates)
-    raise FileNotFoundError(f"Agent profile '{slug}' not found. Searched: {searched}")
-
-
-def _agent_profile_meta(profile: dict) -> dict:
-    """Return the JSON-safe subset stored in tentacle meta.json."""
-    allowed = (
-        "profile_id",
-        "name",
-        "role",
-        "goal",
-        "domain",
-        "expertise",
-        "triggers",
-        "quality_gates",
-        "escalation_rules",
-        "anti_patterns",
-        "evidence_required",
-        "tools_denied",
-        "model",
-        "model_tier",
-        "agent_type",
-        "profile_path",
-    )
-    return {key: profile[key] for key in allowed if key in profile and profile[key] not in ("", [], None)}
-
-
-def _resolve_agent_profile_from_meta(meta: dict) -> dict:
-    """Resolve an embedded or referenced profile from tentacle metadata."""
-    embedded = meta.get("agent_profile")
-    if isinstance(embedded, dict) and embedded.get("profile_id"):
-        profile = dict(embedded)
-        profile["profile_id"] = str(profile["profile_id"])
-        profile["role"] = str(profile.get("role") or profile.get("name") or profile["profile_id"])
-        profile["agent_type"] = str(profile.get("agent_type") or profile["profile_id"])
-        return profile
-    profile_id = meta.get("agent_profile_id")
-    if profile_id:
-        return _load_agent_profile(str(profile_id), find_git_root())
-    return {}
-
-
-def _profile_list(profile: dict, key: str) -> list[str]:
-    """Return a profile field as a clean list of strings."""
-    value = profile.get(key)
-    if isinstance(value, list):
-        return [str(item).strip() for item in value if str(item).strip()]
-    if isinstance(value, str):
-        lines = []
-        for raw_line in value.splitlines():
-            line = raw_line.strip()
-            if line.startswith("- "):
-                line = line[2:].strip()
-            if line:
-                lines.append(line)
-        return lines
-    return []
-
-
-def _render_profile_list(title: str, items: list[str], *, limit: int | None = None) -> list[str]:
-    """Render a titled bullet list when profile data is present."""
-    if not items:
-        return []
-    shown = items[:limit] if limit else items
-    lines = [f"### {title}", ""]
-    lines.extend(f"- {item}" for item in shown)
-    if limit and len(items) > limit:
-        lines.append(f"- ... plus {len(items) - limit} more")
-    lines.append("")
-    return lines
-
-
-def _render_agent_profile_section(profile: dict | None, *, prompt: bool = False) -> str:
-    """Render a specialist profile section for CONTEXT.md or dispatch prompts."""
-    if not profile:
-        return ""
-
-    heading = "### Specialist Profile" if prompt else "## Agent Profile"
-    lines = [heading, ""]
-    profile_id = profile.get("profile_id")
-    role = profile.get("role") or profile.get("name")
-    if profile_id:
-        lines.append(f"**Profile:** `{profile_id}`")
-    if role:
-        lines.append(f"**Role:** {role}")
-    if profile.get("domain"):
-        lines.append(f"**Domain:** {profile['domain']}")
-    if profile.get("model_tier") or profile.get("model"):
-        model_bits = []
-        if profile.get("model_tier"):
-            model_bits.append(f"tier `{profile['model_tier']}`")
-        if profile.get("model"):
-            model_bits.append(f"model `{profile['model']}`")
-        lines.append(f"**Runtime preference:** {', '.join(model_bits)}")
-    if profile.get("goal"):
-        goal = " ".join(str(profile["goal"]).split()) if prompt else str(profile["goal"]).strip()
-        lines.append("")
-        lines.append("**Goal:**")
-        lines.append(goal)
-    lines.append("")
-
-    list_limit = 6 if prompt else None
-    lines.extend(_render_profile_list("Expertise", _profile_list(profile, "expertise"), limit=list_limit))
-    lines.extend(_render_profile_list("Lifecycle Triggers", _profile_list(profile, "triggers"), limit=list_limit))
-    lines.extend(_render_profile_list("Quality Gates", _profile_list(profile, "quality_gates"), limit=list_limit))
-    lines.extend(_render_profile_list("Escalation Rules", _profile_list(profile, "escalation_rules"), limit=list_limit))
-    lines.extend(_render_profile_list("Anti-patterns", _profile_list(profile, "anti_patterns"), limit=list_limit))
-    lines.extend(
-        _render_profile_list(
-            "Evidence Required in Handoff", _profile_list(profile, "evidence_required"), limit=list_limit
-        )
-    )
-    lines.extend(_render_profile_list("Tools Denied", _profile_list(profile, "tools_denied"), limit=list_limit))
-    return "\n" + "\n".join(lines).rstrip() + "\n"
-
-
-def _scope_items(meta: dict) -> list[str]:
-    raw_scope = meta.get("scope") or []
-    if isinstance(raw_scope, str):
-        return [raw_scope.strip()] if raw_scope.strip() else []
-    elif isinstance(raw_scope, list):
-        return [str(item) for item in raw_scope if str(item).strip()]
-    return []
-
-
-def _scope_summary(meta: dict) -> str:
-    items = _scope_items(meta)
-    return ", ".join(items[:6]) if items else "See bundle/session-metadata.md"
-
-
-def _context_excerpt(context: str, limit: int = 180) -> str:
-    lines = [line.strip() for line in context.splitlines() if line.strip()]
-    if not lines:
-        return "See bundle/session-metadata.md"
-    text = " ".join(lines)
-    if len(text) <= limit:
-        return text
-    return text[: limit - 1].rstrip() + "…"
-
-
-def _render_dispatch_context(context: str, meta: dict, bundle_dir: Path | None) -> str:
-    """Render token-lean context for dispatch prompts.
-
-    With a bundle, the file-backed artifact is authoritative; keep the inline
-    prompt small so sub-agents spend tokens on code, not duplicated context.
-    """
-    if not bundle_dir:
-        return context.strip()
-    return textwrap.dedent(f"""\
-        Runtime bundle is authoritative; inline context is intentionally minimal.
-        Read first:
-        1. `{bundle_dir}/context-packet.md`  ← unified context packet
-        2. `{bundle_dir}/manifest.json`
-        3. `{bundle_dir}/session-metadata.md`
-        4. `{bundle_dir}/recall-pack.json`
-        5. `{bundle_dir}/instructions.md` and relevant source files
-
-        Scope: {_scope_summary(meta)}
-        Context excerpt: {_context_excerpt(context)}
-        """).strip()
-
-
-def _render_dispatch_live_briefing_section(briefing_text: str, *, bundled: bool) -> str:
-    """Render the live-briefing section for bundled vs inline dispatch modes."""
-    if not briefing_text:
-        return ""
-    if bundled:
-        return "\n### Live Knowledge\n\nBundled in `briefing.md` and `recall-pack.json`; read those before editing.\n"
-    return f"\n{briefing_text}\n"
-
-
-def _dispatch_context_mode(bundle_dir: Path | None) -> dict:
-    """Describe the active dispatch mode in a JSON-safe structure."""
-    if bundle_dir is not None:
-        return {
-            "name": POINTER_DISPATCH_MODE_NAME,
-            "summary": "Pointer-based bundle (default)",
-            "details": (
-                "Read the Bundle Path files as authoritative context instead of duplicating the "
-                "full tentacle context inline."
-            ),
-            "fallback_flag": "--no-bundle",
-            "default": True,
-        }
-    return {
-        "name": FULL_CONTEXT_DISPATCH_MODE_NAME,
-        "summary": "Full-context inline fallback (`--no-bundle`)",
-        "details": (
-            "This run duplicates the full tentacle context inline because the pointer-based bundle was disabled."
-        ),
-        "fallback_flag": "--no-bundle",
-        "default": False,
-    }
-
-
-def _render_dispatch_mode_section(mode: dict) -> str:
-    """Render a human-readable dispatch-mode section for prompts."""
-    return textwrap.dedent(f"""\
-        ### Dispatch Mode
-
-        **{mode["summary"]}**
-        {mode["details"]}
-        """)
-
-
-def _render_swarm_prompt(
-    name: str,
-    pending: list[dict],
-    context_for_prompt: str,
-    *,
-    specialist_profile_section: str = "",
-    live_briefing_section: str = "",
-    dispatch_mode_section: str = "",
-    prompt_size_section: str = "",
-    bundle_section: str = "",
-    worktree_section: str = "",
-) -> str:
-    """Render the single-agent swarm/dispatch prompt."""
-    prompt = f"""## Tentacle: {name}
-
-### Context
-{context_for_prompt}
-{specialist_profile_section}{live_briefing_section}{dispatch_mode_section}{prompt_size_section}{bundle_section}{worktree_section}
-### Your Tasks (complete ALL)
-"""
-    for t in pending:
-        prompt += f"- [ ] {t['text']}\n"
-
-    prompt += f"""
-### Rules
-- Complete all tasks above
-- If a Bundle Path is present, read `manifest.json` first and use the bundle files as authoritative context
-- Stay within the scoped files only — DO NOT modify files outside your declared scope
-- **DO NOT run `git commit` or `git push`** — the orchestrator owns all git operations
-- **DO NOT widen your scope** beyond the files listed above without explicit escalation to the orchestrator
-- If a task cannot be completed within your scope, stop that task and write a scope escalation note to handoff before continuing
-
-### Cross-review (required before handoff)
-Before writing the handoff, do a self-cross-review:
-1. Re-read every file you modified and confirm correctness against the task description
-2. Verify no unintended changes outside your declared scope
-3. Confirm all todos are complete or explicitly documented as blocked/escalated
-
-### When done
-Mark each completed todo:
-  `python3 ~/.copilot/tools/tentacle.py todo "{name}" done <index>`
-
-Write a structured handoff with status and changed-file receipts:
-  `python3 ~/.copilot/tools/tentacle.py handoff "{name}" "<summary>" --status DONE --changed-file <file1> --changed-file <file2> --learn`
-
-Status values: `DONE` (all tasks complete) | `BLOCKED` (external dependency) | `TOO_BIG` (scope too wide) | `AMBIGUOUS` (spec unclear) | `REGRESSED` (tests broke)
-Add `--changed-file <path>` once per modified file. Omit if no files changed (e.g. BLOCKED with no edits).
-"""
-    return prompt
-
-
-def _dispatch_prompt_size_stats(
-    name: str,
-    pending: list[dict],
-    context: str,
-    meta: dict,
-    *,
-    bundled_live_briefing_section: str,
-    inline_live_briefing_section: str,
-    worktree_section: str,
-    bundle_dir: Path | None,
-    bundle_section: str,
-    agent_profile: dict | None = None,
-) -> dict:
-    """Compare pointer-mode prompt size against the full-context fallback."""
-    specialist_profile_section = _render_agent_profile_section(agent_profile, prompt=True)
-    full_context_prompt = _render_swarm_prompt(
-        name,
-        pending,
-        _render_dispatch_context(context, meta, None),
-        specialist_profile_section=specialist_profile_section,
-        live_briefing_section=inline_live_briefing_section,
-        worktree_section=worktree_section,
-    )
-    if bundle_dir is None:
-        return {
-            "comparison_available": False,
-            "active_prompt_chars": len(full_context_prompt),
-            "full_context_prompt_chars": len(full_context_prompt),
-            "reduction_vs_full_context_percent": 0.0,
-            "minimum_reduction_percent": POINTER_PROMPT_REDUCTION_TARGET_PERCENT,
-            "meets_minimum_reduction": None,
-        }
-
-    pointer_prompt = _render_swarm_prompt(
-        name,
-        pending,
-        _render_dispatch_context(context, meta, bundle_dir),
-        specialist_profile_section=specialist_profile_section,
-        live_briefing_section=bundled_live_briefing_section,
-        bundle_section=bundle_section,
-        worktree_section=worktree_section,
-    )
-    full_chars = max(len(full_context_prompt), 1)
-    reduction = round(max(0.0, (1 - (len(pointer_prompt) / full_chars)) * 100), 1)
-    return {
-        "comparison_available": True,
-        "active_prompt_chars": len(pointer_prompt),
-        "full_context_prompt_chars": len(full_context_prompt),
-        "reduction_vs_full_context_percent": reduction,
-        "minimum_reduction_percent": POINTER_PROMPT_REDUCTION_TARGET_PERCENT,
-        "meets_minimum_reduction": reduction >= POINTER_PROMPT_REDUCTION_TARGET_PERCENT,
-    }
-
-
-def _render_dispatch_prompt_size_section(prompt_size: dict) -> str:
-    """Render prompt-size evidence for dispatch prompts."""
-    if prompt_size.get("comparison_available"):
-        meets = "YES" if prompt_size.get("meets_minimum_reduction") else "NO"
-        return textwrap.dedent(f"""\
-            ### Prompt Size
-
-            - Active prompt chars: `{prompt_size["active_prompt_chars"]}`
-            - Full-context fallback chars: `{prompt_size["full_context_prompt_chars"]}`
-            - Reduction vs full-context: `{prompt_size["reduction_vs_full_context_percent"]:.1f}%` (target: `>= {prompt_size["minimum_reduction_percent"]:.1f}%`)
-            - Meets target: `{meets}`
-            """)
-    return textwrap.dedent(f"""\
-        ### Prompt Size
-
-        - Active prompt chars: `{prompt_size["active_prompt_chars"]}`
-        - Comparison inactive: this run is the full-context inline fallback.
-        """)
+# Dispatch/review helpers live in _tentacle_dispatch.py and _tentacle_review.py.
 
 
 # ---------------------------------------------------------------------------
@@ -1431,776 +749,14 @@ def _get_marker_state() -> dict:
 
 
 # ---------------------------------------------------------------------------
-# SEAM: dispatch-bundle context packet helpers
+# SEAM: dispatch-bundle context packet helpers (extracted to _tentacle_dispatch.py)
 # ---------------------------------------------------------------------------
 
-_DEFAULT_CONTEXT_PACKET_TEMPLATE = """\
-# Context Packet: {{tentacle_name}}
 
-## Tentacle / Scope / Iteration Header
-
-- **Tentacle:** {{tentacle_name}}
-- **Scope:** {{scope}}
-- **Iteration:** {{iteration}}
-- **Status:** {{status}}
-
-## Task Description
-
-{{task_description}}
-
-## Goal Context
-
-{{goal_context}}
-
-## Previous Handoffs
-
-{{previous_handoffs}}
-
-## Project Conventions
-
-{{project_conventions}}
-
-## Recall Pack
-
-{{recall_pack}}
-
-## Blocker Context
-
-{{blocker_context}}
-"""
-
-
-def _apply_context_packet_template(template: str, subs: dict) -> str:
-    """Replace {{key}} markers in *template* with values from *subs*.
-
-    Unknown keys are left as-is so custom templates with extra markers do not
-    crash.  Uses a regex replace to avoid conflicts with Python f-string braces
-    that might appear in literal template content.
-    """
-
-    def _replacer(m: re.Match) -> str:
-        return subs.get(m.group(1), m.group(0))
-
-    return re.sub(r"\{\{(\w+)\}\}", _replacer, template)
-
-
-def _load_context_packet_template(git_root: "Path | None") -> str:
-    """Load the project-level context-packet template if present.
-
-    Looks for ``.github/context-packet-template.md`` under *git_root*.
-    Falls back to ``_DEFAULT_CONTEXT_PACKET_TEMPLATE`` when the file is
-    absent or unreadable.
-    """
-    if git_root:
-        override = git_root / ".github" / "context-packet-template.md"
-        if override.is_file():
-            try:
-                return override.read_text(encoding="utf-8", errors="replace")
-            except OSError:
-                pass
-    return _DEFAULT_CONTEXT_PACKET_TEMPLATE
-
-
-def _blocker_context_from_meta(meta: dict, tentacle_dir: Path) -> str:
-    """Return substantive blocker context when the latest handoff is non-DONE terminal.
-
-    Returns a stable ``None`` placeholder otherwise so the packet is
-    deterministic regardless of run order.
-    """
-    terminal_status = meta.get("terminal_status")
-    if terminal_status not in HANDOFF_TRIAGE_STATUSES:
-        return "None"
-    handoff_path = tentacle_dir / "handoff.md"
-    if not handoff_path.is_file():
-        return f"STATUS: {terminal_status} (see handoff.md — file not present)"
-    try:
-        text = handoff_path.read_text(encoding="utf-8", errors="replace").strip()
-        return text[:600] if len(text) > 600 else text
-    except OSError:
-        return f"STATUS: {terminal_status} (handoff.md unreadable)"
-
-
-def _project_conventions_summary(instr_paths: "list[str]") -> str:
-    """Summarize available instruction sources for the context packet."""
-    if not instr_paths:
-        return "No project instruction files detected.\nSee `bundle/briefing.md` for session-knowledge guidance."
-    shown = instr_paths[:6]
-    lines = [f"- Sources: {', '.join(shown)}"]
-    if len(instr_paths) > len(shown):
-        lines.append(f"- Additional instruction files: {len(instr_paths) - len(shown)}")
-    lines.append("- Full excerpts: `bundle/instructions.md`")
-    lines.append("- Session knowledge: `bundle/briefing.md`")
-    return "\n".join(lines)
-
-
-def _recall_pack_summary(recall_pack_data: "dict | None", recall_source_mode: str | None) -> str:
-    """Summarize the machine-readable recall pack for the context packet."""
-    if not recall_pack_data:
-        return "None — no recall pack data available."
-
-    lines: list[str] = []
-    if recall_source_mode:
-        lines.append(f"- Source mode: {recall_source_mode}")
-
-    tagged = recall_pack_data.get("tagged_entries")
-    if isinstance(tagged, list) and tagged:
-        lines.append(f"- Tagged entries: {len(tagged)}")
-
-    related = recall_pack_data.get("related_entries")
-    if isinstance(related, list) and related:
-        lines.append(f"- Related entries: {len(related)}")
-
-    entries = recall_pack_data.get("entries")
-    if isinstance(entries, dict):
-        populated = []
-        for key, value in entries.items():
-            if isinstance(value, list) and value:
-                populated.append(f"{key}={len(value)}")
-        if populated:
-            lines.append(f"- Entry buckets: {', '.join(populated[:6])}")
-
-    file_matches = recall_pack_data.get("file_matches")
-    if isinstance(file_matches, list) and file_matches:
-        lines.append(f"- File matches: {len(file_matches)}")
-
-    lines.append("- Full payload: `bundle/recall-pack.json`")
-    return "\n".join(lines)
-
-
-def _build_context_packet(
-    name: str,
-    meta: dict,
-    tentacle_dir: Path,
-    context_text: str,
-    goal_context_text: str,
-    prior_handoffs: "list[dict]",
-    recall_pack_data: "dict | None",
-    recall_source_mode: str | None,
-    instr_paths: "list[str]",
-    git_root: "Path | None",
-) -> str:
-    """Render the context-packet.md content for a tentacle bundle.
-
-    Uses the project-level template when present
-    (``.github/context-packet-template.md``), falling back to the built-in
-    default.  All substitution keys use ``{{variable_name}}`` markers so they
-    do not clash with Python f-string syntax.
-    """
-    template = _load_context_packet_template(git_root)
-
-    # ── substitutions ────────────────────────────────────────────────────────
-    scope_str = _scope_summary(meta)
-    task_desc = context_text.strip() or meta.get("description") or "See bundle/session-metadata.md"
-    iteration_str = str(meta.get("goal_iteration") or meta.get("iteration") or "None")
-    status_str = meta.get("status") or "unknown"
-
-    if goal_context_text:
-        goal_section = goal_context_text.strip()
-    else:
-        goal_section = "None — tentacle is not linked to an active goal."
-
-    if prior_handoffs:
-        ph_lines: list[str] = []
-        for entry in prior_handoffs:
-            header = f"[iter-{entry['iteration']} / {entry['tentacle']}]"
-            first_line = entry["summary"].split("\n")[0][:160]
-            ph_lines.append(f"- {header} {first_line}")
-        prev_handoffs_section = "\n".join(ph_lines)
-    else:
-        prev_handoffs_section = "None — first iteration or no prior handoffs recorded."
-
-    project_conventions = _project_conventions_summary(instr_paths)
-    recall_pack_section = _recall_pack_summary(recall_pack_data, recall_source_mode)
-    blocker_section = _blocker_context_from_meta(meta, tentacle_dir)
-
-    subs = {
-        "tentacle_name": name,
-        "scope": scope_str,
-        "iteration": iteration_str,
-        "task_description": task_desc,
-        "status": status_str,
-        "goal_context": goal_section,
-        "previous_handoffs": prev_handoffs_section,
-        "project_conventions": project_conventions,
-        "recall_pack": recall_pack_section,
-        "blocker_context": blocker_section,
-    }
-
-    return _apply_context_packet_template(template, subs)
-
-
-def _build_runtime_bundle(
-    tentacle_dir: Path,
-    name: str,
-    briefing_text: str = "",
-    checkpoint_text: str = "",
-    worktree_path: str | None = None,
-    recall_pack_data: dict | None = None,
-    recall_source_mode: str | None = None,
-    goal_context_text: str = "",
-    context_packet_goal_context_text: str | None = None,
-    prior_handoffs: "list[dict] | None" = None,
-) -> Path:
-    """Materialize a per-run context bundle under the tentacle workspace.
-
-    Creates bundle/ inside the tentacle directory with explicit artifacts:
-      briefing.md         — session-knowledge briefing learnings (or placeholder)
-      instructions.md     — instruction-file surface (host AI config files)
-      skills.md           — skill-file surface (SKILL.md catalogue)
-      session-metadata.md — context, todos, handoff, checkpoint
-      recall-pack.json    — machine-readable recall JSON (task_json or pack mode)
-      manifest.json       — machine-readable index of all artifacts
-
-    Always writes fallback placeholder content for absent surfaces.
-    Returns the bundle directory path.
-    """
-    bundle_dir = tentacle_dir / "bundle"
-    bundle_dir.mkdir(parents=True, exist_ok=True)
-
-    ts = datetime.now(timezone.utc).isoformat()
-    manifest: dict = {
-        "tentacle": name,
-        "created_at": ts,
-        "artifacts": {},
-    }
-    # For collision-renamed tentacles the actual directory name differs from the
-    # logical name; surface it as 'slug' so machine readers can locate the dir.
-    actual_slug = tentacle_dir.name
-    if actual_slug != name:
-        manifest["slug"] = actual_slug
-
-    # ── 1. Briefing ──────────────────────────────────────────────────────────
-    if briefing_text:
-        briefing_content = f"# Briefing: {name}\n\n{briefing_text}\n"
-    else:
-        briefing_content = (
-            f"# Briefing: {name}\n\n"
-            "<!-- No briefing data available for this tentacle. -->\n\n"
-            f'Fetch manually:  python3 ~/.copilot/tools/briefing.py "{name}" --compact\n'
-        )
-    (bundle_dir / "briefing.md").write_text(briefing_content, encoding="utf-8")
-    manifest["artifacts"]["briefing"] = {
-        "file": "briefing.md",
-        "populated": bool(briefing_text),
-    }
-
-    # ── 2. Instruction-file surface ───────────────────────────────────────────
-    instr_lines = ["# Instruction Files\n"]
-    instr_paths: list[str] = []
-    git_root = find_git_root()
-    if git_root:
-        for rel in [
-            ".github/copilot-instructions.md",
-            "CLAUDE.md",
-            "AGENTS.md",
-        ]:
-            p = git_root / rel
-            if p.exists():
-                instr_paths.append(rel)
-                instr_lines.append(f"## {rel}\n")
-                snippet = p.read_text(encoding="utf-8", errors="replace")[:2000]
-                instr_lines.append(snippet)
-                instr_lines.append("\n---\n")
-        instr_dir = git_root / ".github" / "instructions"
-        if instr_dir.exists():
-            for md_file in sorted(instr_dir.glob("*.md")):
-                rel = str(md_file.relative_to(git_root))
-                instr_paths.append(rel)
-                instr_lines.append(f"## {rel}\n")
-                snippet = md_file.read_text(encoding="utf-8", errors="replace")[:1000]
-                instr_lines.append(snippet)
-                instr_lines.append("\n---\n")
-    if not instr_paths:
-        instr_lines.append(
-            "<!-- No instruction files found in this project. -->\n"
-            "Expected: .github/copilot-instructions.md, CLAUDE.md, AGENTS.md, "
-            ".github/instructions/*.md\n"
-        )
-    (bundle_dir / "instructions.md").write_text("\n".join(instr_lines), encoding="utf-8")
-    manifest["artifacts"]["instructions"] = {
-        "file": "instructions.md",
-        "sources": instr_paths,
-        "populated": bool(instr_paths),
-    }
-
-    # ── 3. Skill-file surface ─────────────────────────────────────────────────
-    skill_lines = ["# Skill Files\n"]
-    skill_paths: list[str] = []
-    if git_root:
-        skills_dir = git_root / ".github" / "skills"
-        if skills_dir.exists():
-            for skill_md in sorted(skills_dir.glob("*/SKILL.md")):
-                rel = str(skill_md.relative_to(git_root))
-                skill_name = skill_md.parent.name
-                skill_paths.append(rel)
-                skill_lines.append(f"## {skill_name}\n")
-                snippet = skill_md.read_text(encoding="utf-8", errors="replace")[:500]
-                skill_lines.append(snippet)
-                skill_lines.append("\n---\n")
-    if not skill_paths:
-        skill_lines.append(
-            "<!-- No SKILL.md files found under .github/skills/. -->\n"
-            "Expected pattern: .github/skills/<name>/SKILL.md\n"
-        )
-    (bundle_dir / "skills.md").write_text("\n".join(skill_lines), encoding="utf-8")
-    manifest["artifacts"]["skills"] = {
-        "file": "skills.md",
-        "sources": skill_paths,
-        "populated": bool(skill_paths),
-    }
-
-    # ── 4. Session metadata ───────────────────────────────────────────────────
-    meta_lines = ["# Session Metadata\n"]
-    meta_path = tentacle_dir / "meta.json"
-    context_path = tentacle_dir / "CONTEXT.md"
-    todo_path = tentacle_dir / "todo.md"
-    handoff_path = tentacle_dir / "handoff.md"
-
-    # Load meta once for reuse in the context packet step below.
-    _bundle_meta: dict = {}
-    if meta_path.exists():
-        try:
-            _bundle_meta = json.loads(meta_path.read_text(encoding="utf-8"))
-        except (json.JSONDecodeError, OSError):
-            pass
-
-    if _bundle_meta:
-        meta = _bundle_meta
-        meta_lines.append("## Tentacle Meta\n")
-        meta_lines.append(f"- Name: {meta.get('name', name)}")
-        # Surface the actual directory slug for collision-renamed tentacles so
-        # readers know which directory to look in when name != dir_name.
-        dir_name = meta.get("dir_name")
-        if dir_name:
-            meta_lines.append(f"- Slug: {dir_name}")
-        meta_lines.append(f"- Status: {meta.get('status', 'unknown')}")
-        meta_lines.append(f"- Description: {meta.get('description', '')}")
-        meta_lines.append(f"- Created: {meta.get('created_at', '')}")
-        meta_lines.append("")
-
-    context_text = ""
-    if context_path.exists():
-        context_text = context_path.read_text(encoding="utf-8", errors="replace")
-        meta_lines.append("## Context\n")
-        meta_lines.append(context_text)
-        meta_lines.append("")
-
-    if todo_path.exists():
-        meta_lines.append("## Todos\n")
-        meta_lines.append(todo_path.read_text(encoding="utf-8", errors="replace"))
-        meta_lines.append("")
-
-    if handoff_path.exists():
-        meta_lines.append("## Latest Handoff\n")
-        meta_lines.append(handoff_path.read_text(encoding="utf-8", errors="replace"))
-        meta_lines.append("")
-
-    if checkpoint_text:
-        meta_lines.append("## Checkpoint\n")
-        meta_lines.append(checkpoint_text)
-        meta_lines.append("")
-
-    (bundle_dir / "session-metadata.md").write_text("\n".join(meta_lines), encoding="utf-8")
-    manifest["artifacts"]["session_metadata"] = {
-        "file": "session-metadata.md",
-        "has_context": context_path.exists(),
-        "has_todos": todo_path.exists(),
-        "has_handoff": handoff_path.exists(),
-        "has_checkpoint": bool(checkpoint_text),
-    }
-
-    # ── 5. Recall pack ────────────────────────────────────────────────────────
-    pack_obj: dict = dict(recall_pack_data or {})
-    pack_obj["tentacle"] = name
-    pack_obj["created_at"] = ts
-    pack_obj["source_mode"] = recall_source_mode
-    (bundle_dir / "recall-pack.json").write_text(json.dumps(pack_obj, indent=2) + "\n", encoding="utf-8")
-    manifest["artifacts"]["recall_pack"] = {
-        "file": "recall-pack.json",
-        "populated": bool(recall_pack_data),
-        "source_mode": recall_source_mode,
-    }
-
-    # ── 6. Goal continuation context (optional) ───────────────────────────────
-    if goal_context_text:
-        (bundle_dir / "goal-context.md").write_text(goal_context_text, encoding="utf-8")
-        manifest["artifacts"]["goal_context"] = {
-            "file": "goal-context.md",
-            "populated": True,
-        }
-
-    # ── 7. Context packet (unified agent briefing) ────────────────────────────
-    cp_prior_handoffs: list[dict] = list(prior_handoffs or [])
-    cp_goal_context_text = (
-        context_packet_goal_context_text if context_packet_goal_context_text is not None else goal_context_text
-    )
-    cp_packet = _build_context_packet(
-        name=name,
-        meta=_bundle_meta,
-        tentacle_dir=tentacle_dir,
-        context_text=context_text,
-        goal_context_text=cp_goal_context_text,
-        prior_handoffs=cp_prior_handoffs,
-        recall_pack_data=recall_pack_data,
-        recall_source_mode=recall_source_mode,
-        instr_paths=instr_paths,
-        git_root=git_root,
-    )
-    (bundle_dir / "context-packet.md").write_text(cp_packet, encoding="utf-8")
-    manifest["artifacts"]["context_packet"] = {
-        "file": "context-packet.md",
-        "populated": True,
-        "has_prior_handoffs": bool(cp_prior_handoffs),
-        "has_goal_context": bool(goal_context_text),
-    }
-
-    # ── 8. Manifest ───────────────────────────────────────────────────────────
-    if worktree_path:
-        manifest["worktree_path"] = worktree_path
-    (bundle_dir / "manifest.json").write_text(json.dumps(manifest, indent=2) + "\n", encoding="utf-8")
-
-    return bundle_dir
-
-
-def _reviewer_bundle_dir(tentacle_dir: Path) -> Path:
-    """Return the per-tentacle bundle directory used for fresh-context reviews."""
-    return tentacle_dir / REVIEWER_BUNDLE_DIRNAME
-
-
-def _reviewer_findings_path(tentacle_dir: Path, meta: dict | None = None) -> Path:
-    """Return the persisted reviewer findings path for a tentacle."""
-    reviewer_meta = meta.get("reviewer") if isinstance(meta, dict) else None
-    raw_path = str((reviewer_meta or {}).get("findings_path") or "").strip()
-    if raw_path:
-        return Path(raw_path)
-    return _reviewer_bundle_dir(tentacle_dir) / REVIEWER_FINDINGS_FILENAME
-
-
-def _reviewer_step_paths(name: str, git_root: Path | None) -> list[Path]:
-    """Discover issue/task step files that belong to this tentacle when present."""
-    if not git_root:
-        return []
-    step_dir = git_root / ".github" / "steps"
-    if not step_dir.is_dir():
-        return []
-
-    patterns: list[str] = []
-    issue_match = re.search(r"issue-(\d+)", name)
-    if issue_match:
-        patterns.append(f"issue-{issue_match.group(1)}-*.md")
-    patterns.append(f"{name}.md")
-
-    seen: set[Path] = set()
-    paths: list[Path] = []
-    for pattern in patterns:
-        for path in sorted(step_dir.glob(pattern)):
-            if path.is_file() and path not in seen:
-                seen.add(path)
-                paths.append(path)
-    return paths
-
-
-def _reviewer_render_spec_text(name: str, git_root: Path | None) -> tuple[str, list[str]]:
-    """Render issue-specific step/spec guidance for the reviewer bundle."""
-    paths = _reviewer_step_paths(name, git_root)
-    if not paths:
-        return (
-            "# Reviewer Spec Guidance\n\nNone — no issue-specific step/spec file was found under `.github/steps/`.\n",
-            [],
-        )
-
-    rendered: list[str] = ["# Reviewer Spec Guidance", ""]
-    rel_paths: list[str] = []
-    for path in paths[:3]:
-        rel = str(path.relative_to(git_root)) if git_root else path.name
-        rel_paths.append(rel)
-        rendered.append(f"## {rel}")
-        rendered.append("")
-        rendered.append(path.read_text(encoding="utf-8", errors="replace").strip() or "None")
-        rendered.append("")
-    return "\n".join(rendered).rstrip() + "\n", rel_paths
-
-
-def _reviewer_worktree_path(meta: dict) -> Path:
-    """Resolve the prepared tentacle worktree used as the reviewer diff source."""
-    worktree = meta.get("worktree") or {}
-    worktree_path = str(worktree.get("path") or "").strip()
-    if not worktree.get("prepared") or not worktree_path:
-        raise RuntimeError("dispatch-reviewer requires a prepared tentacle worktree.")
-    path = Path(worktree_path)
-    if not path.exists():
-        raise RuntimeError(f"dispatch-reviewer cannot find the prepared worktree: {worktree_path}")
-    return path
-
-
-def _reviewer_collect_untracked_patch(worktree_path: Path, scope: list[str]) -> str:
-    """Render synthetic unified diffs for untracked files in scope."""
-    cmd = ["git", "-C", str(worktree_path), "ls-files", "--others", "--exclude-standard"]
-    if scope:
-        cmd.extend(["--", *scope])
-    result = subprocess.run(
-        cmd,
-        capture_output=True,
-        text=True,
-        encoding="utf-8",
-        errors="replace",
-    )
-    if result.returncode != 0:
-        raise RuntimeError(result.stderr.strip() or "dispatch-reviewer could not list untracked files.")
-
-    patches: list[str] = []
-    for rel_path in [line.strip() for line in result.stdout.splitlines() if line.strip()]:
-        file_path = worktree_path / rel_path
-        if not file_path.is_file():
-            continue
-        rel_git = rel_path.replace("\\", "/")
-        content = file_path.read_text(encoding="utf-8", errors="replace")
-        diff_lines = list(
-            difflib.unified_diff(
-                [],
-                content.splitlines(),
-                fromfile="/dev/null",
-                tofile=f"b/{rel_git}",
-                lineterm="",
-            )
-        )
-        if not diff_lines:
-            diff_lines = ["--- /dev/null", f"+++ b/{rel_git}"]
-        patches.append(
-            "\n".join(
-                [
-                    f"diff --git a/{rel_git} b/{rel_git}",
-                    "new file mode 100644",
-                    *diff_lines,
-                ]
-            )
-        )
-    return "\n\n".join(patches).strip()
-
-
-def _reviewer_collect_diff(meta: dict) -> tuple[str, str]:
-    """Collect the tentacle-scoped diff shown to the fresh-context reviewer."""
-    worktree_path = _reviewer_worktree_path(meta)
-    scope = _scope_items(meta)
-    cmd = ["git", "-C", str(worktree_path), "diff", "--no-color", "--"]
-    if scope:
-        cmd.extend(scope)
-    result = subprocess.run(
-        cmd,
-        capture_output=True,
-        text=True,
-        encoding="utf-8",
-        errors="replace",
-    )
-    if result.returncode != 0:
-        raise RuntimeError(result.stderr.strip() or "dispatch-reviewer could not capture the scoped diff.")
-
-    diff_text = result.stdout.strip()
-    untracked_patch = _reviewer_collect_untracked_patch(worktree_path, scope)
-    if untracked_patch:
-        diff_text = f"{diff_text}\n\n{untracked_patch}".strip()
-    if not diff_text:
-        raise RuntimeError("dispatch-reviewer found no scoped diff in the prepared worktree.")
-    return diff_text.rstrip() + "\n", str(worktree_path)
-
-
-def _reviewer_findings_template() -> str:
-    """Return the template written for reviewer findings capture."""
-    return textwrap.dedent("""\
-        SAFE_TO_MERGE: PENDING
-
-        ### BLOCKERS
-        - None
-
-        ### WARNINGS
-        - None
-        """)
-
-
-def _extract_markdown_section(text: str, heading: str) -> str | None:
-    """Extract a markdown section body keyed by a ``### <heading>`` marker."""
-    pattern = re.compile(
-        rf"^###\s+{re.escape(heading)}\s*$\n(?P<body>.*?)(?=^###\s+\S|\Z)",
-        flags=re.MULTILINE | re.DOTALL,
-    )
-    match = pattern.search(text or "")
-    if not match:
-        return None
-    return match.group("body").strip()
-
-
-def _reviewer_parse_findings(text: str) -> dict | None:
-    """Parse the structured reviewer output contract from markdown text."""
-    if not text:
-        return None
-    safe_match = re.search(r"^\s*SAFE_TO_MERGE\s*:\s*(.+?)\s*$", text, flags=re.MULTILINE)
-    if not safe_match:
-        return None
-
-    safe_value = safe_match.group(1).strip().upper()
-    if safe_value in REVIEWER_PENDING_VALUES:
-        return None
-    if safe_value not in REVIEWER_SAFE_TRUE | REVIEWER_SAFE_FALSE:
-        return None
-
-    blockers = [
-        item for item in _parse_bullet_list(_extract_markdown_section(text, "BLOCKERS") or "") if item.lower() != "none"
-    ]
-    warnings = [
-        item for item in _parse_bullet_list(_extract_markdown_section(text, "WARNINGS") or "") if item.lower() != "none"
-    ]
-    return {
-        "safe_to_merge": safe_value in REVIEWER_SAFE_TRUE and not blockers,
-        "safe_value": safe_value,
-        "blockers": blockers,
-        "warnings": warnings,
-    }
-
-
-def _reviewer_load_findings(tentacle_dir: Path, meta: dict) -> dict | None:
-    """Load reviewer findings when the structured file has been filled in."""
-    findings_path = _reviewer_findings_path(tentacle_dir, meta)
-    if not findings_path.is_file():
-        return None
-
-    findings_text = findings_path.read_text(encoding="utf-8", errors="replace")
-    parsed = _reviewer_parse_findings(findings_text)
-    if not parsed:
-        return None
-
-    reviewer_meta = meta.get("reviewer") or {}
-    bundle_path = str(reviewer_meta.get("bundle_path") or findings_path.parent)
-    parsed["findings_path"] = str(findings_path)
-    parsed["bundle_path"] = bundle_path
-    parsed["raw_text"] = findings_text
-    return parsed
-
-
-def _build_reviewer_bundle(tentacle_dir: Path, name: str, meta: dict) -> dict:
-    """Materialize the minimal bundle used for fresh-context reviewer dispatch."""
-    bundle_dir = _reviewer_bundle_dir(tentacle_dir)
-    bundle_dir.mkdir(parents=True, exist_ok=True)
-
-    git_root = find_git_root()
-    diff_text, diff_source = _reviewer_collect_diff(meta)
-    spec_text, spec_sources = _reviewer_render_spec_text(name, git_root)
-    scope_lines = [f"- `{path}`" for path in _scope_items(meta)] or ["- None recorded"]
-
-    context_path = tentacle_dir / "CONTEXT.md"
-    context_text = ""
-    if context_path.exists():
-        context_text = context_path.read_text(encoding="utf-8", errors="replace").strip()
-    if not context_text:
-        context_text = meta.get("description") or "No task context recorded."
-
-    review_context = (
-        textwrap.dedent(f"""\
-        # Fresh Context Review: {name}
-
-        Review only the current tentacle diff. Do **not** read implementation handoffs, prior reasoning, or plan files for this review.
-
-        ## Diff Source
-        - `{diff_source}`
-
-        ## Scope
-        {chr(10).join(scope_lines)}
-
-        ## Task Context
-        {context_text}
-
-        ## Review Output Contract
-        - `SAFE_TO_MERGE: YES|NO`
-        - `### BLOCKERS`
-        - `### WARNINGS`
-
-        ## Review-loop Integration
-        - `SAFE_TO_MERGE: YES` and no blockers => the lane may continue.
-        - `BLOCKERS` => review-loop converts them into actionable follow-up context.
-        - `WARNINGS` => preserved but non-blocking.
-        """).rstrip()
-        + "\n"
-    )
-
-    review_context_path = bundle_dir / "review-context.md"
-    review_context_path.write_text(review_context, encoding="utf-8")
-
-    diff_path = bundle_dir / "diff.patch"
-    diff_path.write_text(diff_text, encoding="utf-8")
-
-    spec_path = bundle_dir / "spec.md"
-    spec_path.write_text(spec_text, encoding="utf-8")
-
-    findings_path = bundle_dir / REVIEWER_FINDINGS_FILENAME
-    if not findings_path.exists() or not findings_path.read_text(encoding="utf-8", errors="replace").strip():
-        findings_path.write_text(_reviewer_findings_template(), encoding="utf-8")
-
-    manifest = {
-        "tentacle": name,
-        "created_at": datetime.now(timezone.utc).isoformat(),
-        "diff_source": diff_source,
-        "scope": _scope_items(meta),
-        "artifacts": {
-            "review_context": {"file": review_context_path.name},
-            "diff": {"file": diff_path.name, "populated": True},
-            "spec": {"file": spec_path.name, "sources": spec_sources, "populated": bool(spec_sources)},
-            "findings": {
-                "file": findings_path.name,
-                "populated": _reviewer_parse_findings(findings_path.read_text(encoding="utf-8", errors="replace"))
-                is not None,
-            },
-        },
-    }
-    manifest_path = bundle_dir / "manifest.json"
-    manifest_path.write_text(json.dumps(manifest, indent=2) + "\n", encoding="utf-8")
-
-    return {
-        "bundle_path": str(bundle_dir),
-        "manifest_path": str(manifest_path),
-        "review_context_path": str(review_context_path),
-        "diff_path": str(diff_path),
-        "spec_path": str(spec_path),
-        "findings_path": str(findings_path),
-        "diff_source": diff_source,
-        "spec_sources": spec_sources,
-    }
-
-
-def _render_dispatch_reviewer_prompt(name: str, bundle_info: dict) -> str:
-    """Render the reviewer-specific dispatch prompt."""
-    findings_path = bundle_info["findings_path"]
-    return textwrap.dedent(f"""\
-        ## Fresh-context reviewer: {name}
-
-        Read only these files:
-        1. `{bundle_info["review_context_path"]}`
-        2. `{bundle_info["diff_path"]}`
-        3. `{bundle_info["spec_path"]}`
-        4. `{findings_path}`
-
-        Do **not** read implementation handoffs, prior reasoning, plan files, or full runtime bundles for this review.
-        Do **not** edit code. If your runtime allows file edits, update only `{findings_path}`.
-        If your runtime is read-only, return the exact structure below in chat so the orchestrator can persist it verbatim.
-
-        Return exactly this structure:
-        SAFE_TO_MERGE: YES|NO
-
-        ### BLOCKERS
-        - None
-
-        ### WARNINGS
-        - None
-
-        Review rules:
-        - Surface only genuine bugs, security issues, logic errors, or regression risks.
-        - Use `BLOCKERS` only for must-fix merge blockers.
-        - Use `WARNINGS` for non-blocking concerns.
-        - If there are no blockers, set `SAFE_TO_MERGE: YES`.
-
-        Review-loop integration:
-        - `SAFE_TO_MERGE: YES` and no blockers => the lane may continue.
-        - `BLOCKERS` => review-loop turns them into actionable follow-up context.
-        - `WARNINGS` => visible but non-blocking.
-        """).strip()
+# ---------------------------------------------------------------------------
+# SEAM: review-loop helpers (extracted to _tentacle_review.py)
+# ---------------------------------------------------------------------------
+# Fresh-context review helpers live in _tentacle_review.py.
 
 
 # ---------------------------------------------------------------------------
@@ -2552,346 +1108,7 @@ def cmd_verify(args) -> None:
         sys.exit(exit_code if exit_code > 0 else 1)
 
 
-def _require_done_handoff(tentacle_dir: Path, command_name: str) -> str:
-    """Return the latest handoff content after enforcing a DONE status."""
-    handoff_path = tentacle_dir / "handoff.md"
-    if not handoff_path.exists():
-        print(f"ERROR: {command_name} requires a DONE handoff before running.", file=sys.stderr)
-        sys.exit(1)
-
-    handoff_content = handoff_path.read_text(encoding="utf-8")
-    handoff_status = _parse_handoff_status(handoff_content)
-    if handoff_status != "DONE":
-        print(
-            f"ERROR: {command_name} requires the latest handoff status to be DONE (found: {handoff_status or 'None'}).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    return handoff_content
-
-
-def _review_loop_read_output(log_path: str | None) -> str:
-    """Best-effort read of a verification log file."""
-    if not log_path:
-        return ""
-    try:
-        return Path(log_path).read_text(encoding="utf-8")
-    except OSError:
-        return ""
-
-
-def _review_loop_failure_hash(exit_code: int, output: str) -> str:
-    """Return a stable signature for one failing verification result."""
-    payload = f"{exit_code}:{output}".encode("utf-8", errors="replace")
-    return hashlib.sha256(payload).hexdigest()[:16]
-
-
-def _review_loop_output_excerpt(output: str, *, max_lines: int = 5, max_chars: int = 400) -> str:
-    """Return a short deterministic excerpt for status output and child context."""
-    lines = [line.rstrip() for line in output.splitlines() if line.strip()]
-    if not lines:
-        return "No output captured."
-    excerpt = "\n".join(lines[:max_lines]).strip()
-    if len(excerpt) > max_chars:
-        excerpt = excerpt[: max_chars - 3].rstrip() + "..."
-    return excerpt
-
-
-def _review_loop_latest_verify_command(meta: dict) -> str:
-    """Return the latest recorded verification command, or an empty string."""
-    verifications = list(meta.get("verifications") or [])
-    for record in reversed(verifications):
-        command = str(record.get("command") or "").strip()
-        if command:
-            return command
-    return ""
-
-
-def _review_loop_collect_baseline_failures(meta: dict) -> list[dict]:
-    """Snapshot failing verification signatures that existed before the review loop."""
-    baseline: list[dict] = []
-    seen: set[tuple[str, str]] = set()
-    for record in meta.get("verifications") or []:
-        exit_code = int(record.get("exit_code", 0) or 0)
-        command = str(record.get("command") or "").strip()
-        if exit_code == 0 or not command:
-            continue
-        output = _review_loop_read_output(record.get("log_path"))
-        failure_hash = _review_loop_failure_hash(exit_code, output)
-        key = (command, failure_hash)
-        if key in seen:
-            continue
-        baseline.append({"command": command, "failure_hash": failure_hash})
-        seen.add(key)
-    return baseline
-
-
-def _review_loop_is_preexisting_failure(command: str, failure_hash: str, baseline: list[dict]) -> bool:
-    """Return True when a failure matches the pre-loop baseline for the same command."""
-    for item in baseline:
-        if item.get("command") == command and item.get("failure_hash") == failure_hash:
-            return True
-    return False
-
-
-def _review_loop_is_test_file(path: str) -> bool:
-    """Return True when a path looks like a test file."""
-    normalized = path.replace("\\", "/").lstrip("./").lower()
-    name = normalized.rsplit("/", 1)[-1]
-    return (
-        normalized.startswith("tests/")
-        or "/tests/" in normalized
-        or name.startswith("test_")
-        or name.endswith("_test.py")
-    )
-
-
-def _review_loop_changed_files(tentacle_dir: Path, meta: dict) -> list[str]:
-    """Return changed files from meta and the latest handoff entry."""
-    changed: list[str] = []
-    for path in meta.get("changed_files") or []:
-        candidate = str(path).strip()
-        if candidate and candidate not in changed:
-            changed.append(candidate)
-    handoff_path = tentacle_dir / "handoff.md"
-    if not handoff_path.exists():
-        return changed
-    raw = handoff_path.read_text(encoding="utf-8")
-    for path in _parse_handoff_changed_files(raw):
-        if path and path not in changed:
-            changed.append(path)
-    rich = _parse_rich_handoff_sections(raw)
-    for path in rich.get("files_modified") or []:
-        if path and path not in changed:
-            changed.append(path)
-    return changed
-
-
-def _review_loop_classify_failure(command: str, output: str, changed_files: list[str]) -> tuple[str, str]:
-    """Classify a new actionable review-loop failure."""
-    command_lower = command.lower()
-    for token in REVIEW_LOOP_BUILD_ERROR_COMMAND_TOKENS:
-        if token in command_lower:
-            return "BUILD_ERROR", f"verification command targets build tooling ({token})"
-    for pattern in REVIEW_LOOP_BUILD_ERROR_PATTERNS:
-        if re.search(pattern, output, flags=re.IGNORECASE):
-            return "BUILD_ERROR", f"verification output matched build-error pattern: {pattern}"
-    if changed_files and all(_review_loop_is_test_file(path) for path in changed_files):
-        for pattern in REVIEW_LOOP_TEST_FAILURE_PATTERNS:
-            if re.search(pattern, output, flags=re.IGNORECASE):
-                return "NEW_TEST_WRONG", "only test files changed and the failure looks like a test expectation issue"
-    return "REGRESSION", "new failure does not match the pre-existing, flaky, build-error, or test-only buckets"
-
-
-def _review_loop_history(meta: dict) -> list[dict]:
-    """Return persisted review-loop history records."""
-    review_loop = meta.get("review_loop") or {}
-    history = review_loop.get("history") or []
-    return history if isinstance(history, list) else []
-
-
-def _review_loop_actionable_count(meta: dict) -> int:
-    """Count prior actionable review-loop iterations."""
-    return sum(
-        1
-        for entry in _review_loop_history(meta)
-        if str(entry.get("classification") or "") in REVIEW_LOOP_ACTIONABLE_CLASSIFICATIONS
-    )
-
-
-def _review_loop_next_resolver_name(parent_name: str, tentacles: Path) -> str:
-    """Return a unique short child tentacle name for a blocker resolver."""
-    base = _tentacle_slug(parent_name)[:48].rstrip("-") or "tentacle"
-    index = 1
-    while True:
-        candidate = f"{base}-review-fix-{index}"
-        if not (tentacles / candidate).exists():
-            return candidate
-        index += 1
-
-
-def _review_loop_append_history(
-    meta_path: Path,
-    *,
-    baseline_failures: list[dict],
-    entry: dict,
-    unresolved_blockers: list[str] | None = None,
-) -> dict:
-    """Append one review-loop history record and persist the baseline snapshot."""
-    meta = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    review_loop = meta.setdefault("review_loop", {})
-    history = list(review_loop.get("history") or [])
-    stored_entry = dict(entry)
-    stored_entry["index"] = len(history) + 1
-    stored_entry["recorded_at"] = datetime.now(timezone.utc).isoformat()
-    history.append(stored_entry)
-    review_loop["history"] = history
-    review_loop["baseline_failures"] = baseline_failures
-    review_loop["last_classification"] = stored_entry.get("classification")
-    review_loop["updated_at"] = datetime.now(timezone.utc).isoformat()
-    if unresolved_blockers:
-        review_loop["unresolved_blockers"] = unresolved_blockers
-    else:
-        review_loop.pop("unresolved_blockers", None)
-    meta["review_loop"] = review_loop
-    meta_path.write_text(json.dumps(meta, indent=2) + "\n", encoding="utf-8")
-    return stored_entry
-
-
-def _review_loop_create_resolver_tentacle(
-    *,
-    args,
-    tentacles: Path,
-    parent_name: str,
-    parent_dir: Path,
-    parent_meta: dict,
-    classification: str,
-    verification_command: str,
-    failure_excerpt: str,
-    actionable_iteration: int,
-    reviewer_findings: dict | None = None,
-) -> str:
-    """Create a follow-up blocker-resolver tentacle that inherits parent scope/worktree."""
-    resolver_name = _review_loop_next_resolver_name(parent_name, tentacles)
-    resolver_dir = tentacles / resolver_name
-    resolver_dir.mkdir(parents=True, exist_ok=False)
-
-    parent_scope = _scope_items(parent_meta)
-    desc = f"Resolve {classification} from review-loop for {parent_name}"
-    context_lines = [
-        f"# {resolver_name}",
-        "",
-        desc,
-        "",
-        "## Parent Tentacle",
-        "",
-        f"- `{parent_name}`",
-        f"- Review-loop iteration: `{actionable_iteration}`",
-        f"- Classification: `{classification}`",
-        f"- Verification command: `{verification_command}`",
-    ]
-    worktree = parent_meta.get("worktree") or {}
-    worktree_path = str(worktree.get("path") or "").strip()
-    if worktree.get("prepared") and worktree_path:
-        context_lines.extend(
-            ["- Reuse inherited worktree path below; do not prepare a new worktree.", f"- `{worktree_path}`"]
-        )
-    if parent_scope:
-        context_lines.extend(["", "## Scope", ""])
-        context_lines.extend([f"- `{path}`" for path in parent_scope])
-    if reviewer_findings and reviewer_findings.get("blockers"):
-        context_lines.extend(["", "## Reviewer Blockers", ""])
-        context_lines.extend([f"- {item}" for item in reviewer_findings["blockers"]])
-    if reviewer_findings and reviewer_findings.get("warnings"):
-        context_lines.extend(["", "## Reviewer Warnings", ""])
-        context_lines.extend([f"- {item}" for item in reviewer_findings["warnings"]])
-    context_lines.extend(
-        [
-            "",
-            "## Failure Excerpt",
-            "",
-            "```text",
-            failure_excerpt,
-            "```",
-            "",
-            "## Constraints",
-            "",
-            "- Stay in the inherited parent scope and worktree.",
-            "- Fix the review-loop failure without widening into unrelated orchestration changes.",
-            "- Rerun the same verification command before writing handoff.",
-            "",
-            "## Key files",
-            "",
-        ]
-    )
-    if parent_scope:
-        context_lines.extend([f"- `{path}`" for path in parent_scope])
-    else:
-        context_lines.append("- Reuse the parent tentacle scope")
-    context_lines.extend(["", "---", f"*Created: {datetime.now(timezone.utc).isoformat()}*"])
-    (resolver_dir / "CONTEXT.md").write_text("\n".join(context_lines) + "\n", encoding="utf-8")
-
-    todos = [
-        {
-            "index": 0,
-            "done": False,
-            "text": f"Diagnose the {classification} review-loop failure from `{parent_name}`.",
-            "line_number": 0,
-        },
-        {
-            "index": 1,
-            "done": False,
-            "text": f"Fix the failure in the inherited worktree and rerun `{verification_command}`.",
-            "line_number": 0,
-        },
-        {
-            "index": 2,
-            "done": False,
-            "text": "Write a structured handoff with changed-file receipts and the rerun result.",
-            "line_number": 0,
-        },
-    ]
-    (resolver_dir / "todo.md").write_text(render_todos(todos), encoding="utf-8")
-
-    resolver_meta = {
-        "name": resolver_name,
-        "created_at": datetime.now(timezone.utc).isoformat(),
-        "scope": parent_scope,
-        "description": desc,
-        "status": "idle",
-        "tentacle_id": str(uuid.uuid4()),
-        "skills": [],
-        "review_loop_parent": parent_name,
-        "review_loop_classification": classification,
-        "review_loop_verification_command": verification_command,
-        "review_loop_failure_excerpt": failure_excerpt,
-        "review_loop_iteration": actionable_iteration,
-    }
-    if reviewer_findings:
-        resolver_meta["reviewer_blockers"] = list(reviewer_findings.get("blockers") or [])
-        resolver_meta["reviewer_warnings"] = list(reviewer_findings.get("warnings") or [])
-        if reviewer_findings.get("findings_path"):
-            resolver_meta["reviewer_findings_path"] = reviewer_findings["findings_path"]
-    goal_id = parent_meta.get("goal_id")
-    iteration = parent_meta.get("goal_iteration") or parent_meta.get("iteration")
-    if goal_id:
-        resolver_meta["goal_id"] = goal_id
-    if iteration is not None:
-        resolver_meta["goal_iteration"] = iteration
-        resolver_meta["iteration"] = iteration
-    (resolver_dir / "meta.json").write_text(json.dumps(resolver_meta, indent=2) + "\n", encoding="utf-8")
-
-    if worktree.get("prepared") and worktree_path and Path(worktree_path).exists():
-        inherited_state = dict(worktree)
-        inherited_state["reused"] = True
-        _update_meta_worktree(resolver_dir, inherited_state)
-
-    if goal_id:
-        try:
-            goal_state = _goal_load(tentacles)
-            if goal_state and goal_state.get("goal_id") == goal_id:
-                _cmd_goal_link(argparse.Namespace(tentacle_name=resolver_name), tentacles)
-        except SystemExit:
-            raise
-        except Exception:
-            pass
-
-    return resolver_name
-
-
-def _review_loop_dispatch_resolver(args, resolver_name: str) -> None:
-    """Emit the standard swarm/dispatch prompt for a blocker-resolver tentacle."""
-    dispatch_args = argparse.Namespace(
-        session_dir=args.session_dir,
-        name=resolver_name,
-        agent_type=getattr(args, "agent_type", None) or "general-purpose",
-        model=getattr(args, "model", None) or "claude-sonnet-4.6",
-        output="prompt",
-        briefing=False,
-        bundle=True,
-        worktree=False,
-    )
-    cmd_swarm(dispatch_args)
+# Review-loop helpers live in _tentacle_review.py.
 
 
 def _normalize_changed_file_paths(paths: "list[str] | None") -> list[str]:
@@ -3153,274 +1370,7 @@ def _apply_scope_reclassification(
     return record
 
 
-def _review_loop_handle_reviewer_findings(
-    *,
-    args,
-    tentacles: Path,
-    tentacle_dir: Path,
-    meta_path: Path,
-    meta_before: dict,
-    baseline_failures: list[dict],
-    history_before: list[dict],
-    verification_command: str,
-    first_run: dict,
-    existing_actionable: int,
-    max_iterations: int,
-) -> bool:
-    """Consume fresh-context reviewer findings when a populated result is available."""
-    reviewer_findings = _reviewer_load_findings(tentacle_dir, meta_before)
-    if not reviewer_findings:
-        return False
-
-    warnings = list(reviewer_findings.get("warnings") or [])
-    blockers = list(reviewer_findings.get("blockers") or [])
-    if not blockers and reviewer_findings.get("safe_to_merge"):
-        entry = {
-            "command": verification_command,
-            "classification": "PASS",
-            "outcome": "passed",
-            "reason": "fresh-context reviewer marked the diff safe to merge",
-            "verify_runs": [first_run],
-            "history_size_before": len(history_before),
-            "reviewer_findings": reviewer_findings,
-        }
-        _review_loop_append_history(meta_path, baseline_failures=baseline_failures, entry=entry)
-        print("✅ review-loop passed verification and fresh-context review.")
-        if warnings:
-            print("\n### REVIEWER WARNINGS")
-            for item in warnings:
-                print(f"- {item}")
-        return True
-
-    actionable_iteration = existing_actionable + 1
-    failure_excerpt = (
-        "\n".join(blockers) if blockers else "Reviewer marked SAFE_TO_MERGE: NO without explicit blockers."
-    )
-    entry = {
-        "command": verification_command,
-        "classification": "REGRESSION",
-        "reason": "fresh-context reviewer reported blockers",
-        "outcome": "resolver_created",
-        "verify_runs": [first_run],
-        "changed_files": _review_loop_changed_files(tentacle_dir, meta_before),
-        "actionable_iteration": actionable_iteration,
-        "history_size_before": len(history_before),
-        "reviewer_findings": reviewer_findings,
-    }
-
-    if existing_actionable >= max_iterations:
-        unresolved_blockers = blockers or [failure_excerpt]
-        entry["outcome"] = "unresolved"
-        entry["resolver_tentacle"] = None
-        _review_loop_append_history(
-            meta_path,
-            baseline_failures=baseline_failures,
-            entry=entry,
-            unresolved_blockers=unresolved_blockers,
-        )
-        print("❌ review-loop retry budget exhausted on reviewer blockers.")
-        print("\n### UNRESOLVED BLOCKERS")
-        for blocker in unresolved_blockers:
-            print(f"- {blocker}")
-        sys.exit(1)
-
-    resolver_name = _review_loop_create_resolver_tentacle(
-        args=args,
-        tentacles=tentacles,
-        parent_name=args.name,
-        parent_dir=tentacle_dir,
-        parent_meta=meta_before,
-        classification="REGRESSION",
-        verification_command=verification_command,
-        failure_excerpt=failure_excerpt,
-        actionable_iteration=actionable_iteration,
-        reviewer_findings=reviewer_findings,
-    )
-    entry["resolver_tentacle"] = resolver_name
-    _review_loop_append_history(meta_path, baseline_failures=baseline_failures, entry=entry)
-    print("🛠️  review-loop received BLOCKERS from the fresh-context reviewer.")
-    print(f"   Resolver tentacle: {resolver_name}")
-    if warnings:
-        print("\n### REVIEWER WARNINGS")
-        for item in warnings:
-            print(f"- {item}")
-    print("")
-    _review_loop_dispatch_resolver(args, resolver_name)
-    sys.exit(1)
-
-
-def cmd_review_loop(args) -> None:
-    """Run a tentacle-scoped self-healing review loop."""
-    tentacles = get_tentacles_dir(args.session_dir)
-    tentacle_dir = _validate_tentacle_name(args.name, tentacles)
-    if not tentacle_dir.exists():
-        print(f"ERROR: Tentacle '{args.name}' not found.", file=sys.stderr)
-        sys.exit(1)
-
-    _require_done_handoff(tentacle_dir, "review-loop")
-
-    meta_path = tentacle_dir / "meta.json"
-    meta_before = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    verification_command = (getattr(args, "verify_command", None) or "").strip()
-    if not verification_command:
-        verification_command = _review_loop_latest_verify_command(meta_before)
-    if not verification_command:
-        print(
-            "ERROR: No verification command supplied and no prior verification record exists for this tentacle.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    timeout = getattr(args, "timeout", 120) or 120
-    max_iterations = getattr(args, "max_iterations", 5) or 5
-    history_before = _review_loop_history(meta_before)
-    review_loop_state = meta_before.get("review_loop") or {}
-    baseline_failures = review_loop_state.get("baseline_failures")
-    if not isinstance(baseline_failures, list):
-        baseline_failures = _review_loop_collect_baseline_failures(meta_before)
-    existing_actionable = _review_loop_actionable_count(meta_before)
-
-    print(f"🔄 review-loop: '{args.name}'")
-    print(f"   verify: {verification_command}")
-    print(f"   max-iterations: {max_iterations}")
-
-    first_exit, first_record = _run_and_record_verification(
-        tentacle_dir=tentacle_dir,
-        meta=meta_before,
-        meta_path=meta_path,
-        cmd=verification_command,
-        label="review-loop",
-        timeout=timeout,
-    )
-    first_output = _review_loop_read_output(first_record.get("log_path"))
-    first_hash = _review_loop_failure_hash(first_exit, first_output)
-    first_run = {
-        "exit_code": first_exit,
-        "log_path": first_record.get("log_path"),
-        "failure_hash": first_hash,
-        "output_excerpt": _review_loop_output_excerpt(first_output),
-    }
-
-    if first_exit == 0:
-        if _review_loop_handle_reviewer_findings(
-            args=args,
-            tentacles=tentacles,
-            tentacle_dir=tentacle_dir,
-            meta_path=meta_path,
-            meta_before=meta_before,
-            baseline_failures=baseline_failures,
-            history_before=history_before,
-            verification_command=verification_command,
-            first_run=first_run,
-            existing_actionable=existing_actionable,
-            max_iterations=max_iterations,
-        ):
-            return
-        entry = {
-            "command": verification_command,
-            "classification": "PASS",
-            "outcome": "passed",
-            "verify_runs": [first_run],
-            "history_size_before": len(history_before),
-        }
-        _review_loop_append_history(meta_path, baseline_failures=baseline_failures, entry=entry)
-        print("✅ review-loop passed on the first verification run.")
-        return
-
-    if _review_loop_is_preexisting_failure(verification_command, first_hash, baseline_failures):
-        entry = {
-            "command": verification_command,
-            "classification": "PRE_EXISTING",
-            "outcome": "ignored",
-            "reason": "failure matches the pre-loop verification baseline",
-            "verify_runs": [first_run],
-            "history_size_before": len(history_before),
-        }
-        _review_loop_append_history(meta_path, baseline_failures=baseline_failures, entry=entry)
-        print("⚪ review-loop classified the failure as PRE_EXISTING — no blocker-resolver created.")
-        return
-
-    meta_retry = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    second_exit, second_record = _run_and_record_verification(
-        tentacle_dir=tentacle_dir,
-        meta=meta_retry,
-        meta_path=meta_path,
-        cmd=verification_command,
-        label="review-loop-retry",
-        timeout=timeout,
-    )
-    second_output = _review_loop_read_output(second_record.get("log_path"))
-    second_run = {
-        "exit_code": second_exit,
-        "log_path": second_record.get("log_path"),
-        "failure_hash": _review_loop_failure_hash(second_exit, second_output),
-        "output_excerpt": _review_loop_output_excerpt(second_output),
-    }
-
-    if second_exit == 0:
-        entry = {
-            "command": verification_command,
-            "classification": "FLAKY",
-            "outcome": "passed",
-            "reason": "initial failure disappeared on the immediate retry",
-            "verify_runs": [first_run, second_run],
-            "history_size_before": len(history_before),
-        }
-        _review_loop_append_history(meta_path, baseline_failures=baseline_failures, entry=entry)
-        print("🟡 review-loop classified the failure as FLAKY — retry passed, no blocker-resolver created.")
-        return
-
-    meta_after_retries = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    changed_files = _review_loop_changed_files(tentacle_dir, meta_after_retries)
-    classification, reason = _review_loop_classify_failure(verification_command, second_output, changed_files)
-    failure_excerpt = _review_loop_output_excerpt(second_output)
-    actionable_iteration = existing_actionable + 1
-    entry = {
-        "command": verification_command,
-        "classification": classification,
-        "reason": reason,
-        "outcome": "resolver_created",
-        "verify_runs": [first_run, second_run],
-        "changed_files": changed_files,
-        "actionable_iteration": actionable_iteration,
-        "history_size_before": len(history_before),
-    }
-
-    if existing_actionable >= max_iterations:
-        blockers = [f"{classification}: {failure_excerpt}"]
-        entry["outcome"] = "unresolved"
-        entry["resolver_tentacle"] = None
-        _review_loop_append_history(
-            meta_path,
-            baseline_failures=baseline_failures,
-            entry=entry,
-            unresolved_blockers=blockers,
-        )
-        print("❌ review-loop retry budget exhausted.")
-        print("\n### UNRESOLVED BLOCKERS")
-        for blocker in blockers:
-            print(f"- {blocker}")
-        sys.exit(1)
-
-    resolver_name = _review_loop_create_resolver_tentacle(
-        args=args,
-        tentacles=tentacles,
-        parent_name=args.name,
-        parent_dir=tentacle_dir,
-        parent_meta=meta_after_retries,
-        classification=classification,
-        verification_command=verification_command,
-        failure_excerpt=failure_excerpt,
-        actionable_iteration=actionable_iteration,
-    )
-    entry["resolver_tentacle"] = resolver_name
-    _review_loop_append_history(meta_path, baseline_failures=baseline_failures, entry=entry)
-    print(f"🛠️  review-loop classified the failure as {classification}.")
-    print(f"   Reason: {reason}")
-    print(f"   Resolver tentacle: {resolver_name}")
-    print("")
-    _review_loop_dispatch_resolver(args, resolver_name)
-    sys.exit(1)
+# Review-loop command orchestration lives in _tentacle_review.py.
 
 
 # ---------------------------------------------------------------------------
@@ -5133,543 +3083,13 @@ def cmd_audit(args):
             print(f"✅ Audit clean for '{args.name}': all changed/scoped files were read.")
 
 
-def cmd_resume(args):
-    """Resume a tentacle: refresh briefing, update status, and show current state."""
-    tentacles = get_tentacles_dir(args.session_dir)
-    tentacle_dir = _validate_tentacle_name(args.name, tentacles)
-
-    if not tentacle_dir.exists():
-        print(f"ERROR: Tentacle '{args.name}' not found.", file=sys.stderr)
-        sys.exit(1)
-
-    meta_path = tentacle_dir / "meta.json"
-    context_path = tentacle_dir / "CONTEXT.md"
-    todo_path = tentacle_dir / "todo.md"
-    handoff_path = tentacle_dir / "handoff.md"
-
-    # 1. Load and update meta
-    meta = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    prev_status = meta.get("status", "idle")
-    meta["status"] = "active"
-    meta["resumed_at"] = datetime.now(timezone.utc).isoformat()
-    meta_path.write_text(json.dumps(meta, indent=2) + "\n", encoding="utf-8")
-
-    print(f"🔄 Resuming tentacle '{args.name}' (was: {prev_status})")
-
-    # 2. Live briefing injection (unless --no-briefing)
-    briefing_text = ""
-    checkpoint_text = ""
-    if not getattr(args, "no_briefing", False):
-        fallback = meta.get("description", "") or args.name.replace("-", " ")
-        print(f"🧠 Fetching fresh knowledge for '{args.name}'...")
-        briefing_text = _run_briefing_for_task(args.name, fallback_query=fallback)
-        if briefing_text:
-            print(f"   ✅ Got {len(briefing_text)} chars of relevant knowledge")
-        else:
-            print("   ℹ️  No relevant past knowledge found")
-        checkpoint_text = _load_latest_checkpoint_context()
-        if checkpoint_text:
-            print("   📌 Latest checkpoint context injected")
-
-    # 3. Replace a bounded AUTO-RECALL block in CONTEXT.md
-    timestamp = datetime.now(timezone.utc).strftime("%Y-%m-%d %H:%M UTC")
-    recall_lines = [f"## Resumed [{timestamp}]"]
-    if briefing_text:
-        recall_lines.append(briefing_text)
-    else:
-        recall_lines.append("_No new briefing content available._")
-    if checkpoint_text:
-        recall_lines.append(checkpoint_text)
-    recall_content = "\n\n".join(recall_lines).rstrip()
-    if context_path.exists():
-        existing = context_path.read_text(encoding="utf-8")
-        updated = _upsert_auto_recall_block(existing, recall_content)
-        context_path.write_text(updated, encoding="utf-8")
-    else:
-        base_context = f"# {args.name}\n"
-        updated = _upsert_auto_recall_block(base_context, recall_content)
-        context_path.write_text(updated, encoding="utf-8")
-
-    # 4. Show current todo state
-    todos = parse_todos(todo_path.read_text(encoding="utf-8")) if todo_path.exists() else []
-    done_count = sum(1 for t in todos if t["done"])
-    pending = [t for t in todos if not t["done"]]
-
-    print(f"\n📋 Todos: {done_count}/{len(todos)} done")
-    if pending:
-        print("   Pending:")
-        for t in pending:
-            print(f"     ☐ [{t['index']}] {t['text']}")
-    else:
-        print("   ✅ All todos done" if todos else "   (none yet)")
-
-    if handoff_path.exists():
-        print(f"\n📨 Handoff notes available — run `show {args.name}` to review")
-
-    print(f"\n✅ Tentacle '{args.name}' is active and ready")
+# Dispatch commands live in _tentacle_dispatch.py.
 
 
-def cmd_swarm(args):
-    """Generate dispatch instructions from pending todos (swarm mode)."""
-    tentacles = get_tentacles_dir(args.session_dir)
-    tentacle_dir = _validate_tentacle_name(args.name, tentacles)
-
-    if not tentacle_dir.exists():
-        print(f"ERROR: Tentacle '{args.name}' not found.", file=sys.stderr)
-        sys.exit(1)
-
-    todo_path = tentacle_dir / "todo.md"
-    context_path = tentacle_dir / "CONTEXT.md"
-    meta_path = tentacle_dir / "meta.json"
-
-    todos = parse_todos(todo_path.read_text(encoding="utf-8")) if todo_path.exists() else []
-    pending = [t for t in todos if not t["done"]]
-
-    if not pending:
-        print(f"✅ All todos done for '{args.name}'. Nothing to swarm.")
-        return
-
-    bundle_enabled = _bundle_enabled(args)
-
-    if args.output == "json" and getattr(args, "briefing", False) and not bundle_enabled:
-        print(
-            "ERROR: --briefing is not supported with --output json. "
-            "Use the default runtime bundle (or pass --bundle) so briefing "
-            "can be represented via recall-pack.json, or use --output prompt/parallel.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    context = context_path.read_text(encoding="utf-8") if context_path.exists() else ""
-    meta = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-
-    try:
-        agent_profile = _resolve_agent_profile_from_meta(meta)
-    except (FileNotFoundError, ValueError) as exc:
-        print(f"ERROR: {exc}", file=sys.stderr)
-        sys.exit(1)
-
-    requested_agent_type = getattr(args, "agent_type", None)
-    requested_model = getattr(args, "model", None)
-    agent_type = (
-        requested_agent_type
-        or meta.get("agent_type")
-        or agent_profile.get("agent_type")
-        or agent_profile.get("profile_id")
-        or "general-purpose"
-    )
-    model = requested_model or meta.get("model") or agent_profile.get("model") or "claude-sonnet-4.6"
-    specialist_profile_section = _render_agent_profile_section(agent_profile, prompt=True)
-
-    print(f"🐙 Swarm plan for '{args.name}' — {len(pending)} pending todos\n")
-    print(f"Agent: {agent_type} | Model: {model}\n")
-    if agent_profile:
-        print(f"Profile: {agent_profile.get('profile_id')} | Role: {agent_profile.get('role')}\n")
-
-    # Live briefing injection at dispatch time
-    briefing_text = ""
-    briefing_recall_data: dict = {}
-    briefing_recall_mode: str | None = None
-    if getattr(args, "briefing", False):
-        fallback = meta.get("description", "") or args.name.replace("-", " ")
-        print("🧠 Fetching live briefing for dispatch...")
-        if bundle_enabled:
-            briefing_recall_data, briefing_recall_mode = _fetch_recall_pack_json(
-                args.name,
-                fallback_query=fallback,
-            )
-            briefing_text = _render_recall_payload(
-                args.name,
-                briefing_recall_data,
-                briefing_recall_mode,
-            )
-        else:
-            briefing_text = _run_briefing_for_task(args.name, fallback_query=fallback)
-        if briefing_text:
-            print(f"   ✅ Injected {len(briefing_text)} chars of live knowledge\n")
-        else:
-            print("   ℹ️  No relevant past knowledge found\n")
-    bundled_live_briefing_section = _render_dispatch_live_briefing_section(briefing_text, bundled=True)
-    inline_live_briefing_section = _render_dispatch_live_briefing_section(briefing_text, bundled=False)
-    active_live_briefing_section = bundled_live_briefing_section if bundle_enabled else inline_live_briefing_section
-
-    # Bundle materialization (default for CLI swarm/dispatch; opt out with --no-bundle)
-    bundle_dir: Path | None = None
-    bundle_section = ""
-    worktree_section = ""
-
-    # Worktree preparation (--worktree flag)
-    wt_path_str: str | None = None
-    if getattr(args, "worktree", False):
-        print(f"🌿 Preparing worktree for '{args.name}'...")
-        git_root = find_git_root()
-        wt_state = _worktree_prepare(tentacle_dir, args.name, git_root)
-        if wt_state["prepared"]:
-            wt_path_str = wt_state["path"]
-            action = "reused" if wt_state.get("reused") else "prepared"
-            print(f"   ✅ Worktree {action}: {wt_path_str}\n")
-            worktree_section = f"\n### Worktree Path\n\n`{wt_path_str}`\n"
-        else:
-            print(f"   ⚠️  Worktree prepare failed: {wt_state.get('error', 'unknown')}\n")
-
-    if bundle_enabled:
-        print("📦 Materializing runtime bundle...")
-        b_fallback = meta.get("description", "") or args.name.replace("-", " ")
-        b_checkpoint = _load_latest_checkpoint_context()
-        if getattr(args, "briefing", False):
-            b_recall, b_recall_mode = briefing_recall_data, briefing_recall_mode
-            b_briefing = briefing_text
-        else:
-            b_recall, b_recall_mode = _fetch_recall_pack_json(args.name, fallback_query=b_fallback)
-            b_briefing = _render_recall_payload(args.name, b_recall, b_recall_mode)
-        # Gather goal context if this tentacle is linked to an active goal
-        swarm_goal_context_text = ""
-        swarm_packet_goal_context_text = ""
-        swarm_prior_handoffs: list[dict] = []
-        try:
-            swarm_goal_state = _goal_load(tentacles)
-            if swarm_goal_state and args.name in (swarm_goal_state.get("tentacles") or []):
-                swarm_goal_context_text = _goal_render_continuation_context(swarm_goal_state, tentacles)
-                swarm_packet_goal_context_text = _goal_render_continuation_context(
-                    swarm_goal_state,
-                    tentacles,
-                    include_prior_handoffs=False,
-                )
-                swarm_prior_handoffs = _goal_collect_prior_handoffs(swarm_goal_state, tentacles)
-        except Exception:
-            pass
-        bundle_dir = _build_runtime_bundle(
-            tentacle_dir=tentacle_dir,
-            name=args.name,
-            briefing_text=b_briefing,
-            checkpoint_text=b_checkpoint,
-            worktree_path=wt_path_str,
-            recall_pack_data=b_recall,
-            recall_source_mode=b_recall_mode,
-            goal_context_text=swarm_goal_context_text,
-            context_packet_goal_context_text=swarm_packet_goal_context_text,
-            prior_handoffs=swarm_prior_handoffs,
-        )
-        bundle_section = (
-            "\n### Bundle Path\n\n"
-            f"`{bundle_dir}`\n\n"
-            "Use this bundle as the source of truth for full context; do not duplicate it into the prompt.\n"
-        )
-        print(f"   ✅ Bundle: {bundle_dir}\n")
-
-    dispatch_context_mode = _dispatch_context_mode(bundle_dir)
-    prompt_size = _dispatch_prompt_size_stats(
-        args.name,
-        pending,
-        context,
-        meta,
-        bundled_live_briefing_section=bundled_live_briefing_section,
-        inline_live_briefing_section=inline_live_briefing_section,
-        worktree_section=worktree_section,
-        bundle_dir=bundle_dir,
-        bundle_section=bundle_section,
-        agent_profile=agent_profile,
-    )
-    dispatch_mode_section = _render_dispatch_mode_section(dispatch_context_mode)
-    prompt_size_section = _render_dispatch_prompt_size_section(prompt_size)
-
-    # Write dispatched-subagent-active marker so local enforcement surfaces can
-    # observe that a dispatch is in flight. The marker is advisory — tentacle.py
-    # is not itself an enforcement layer. Cleared by cmd_complete.
-    tentacle_id = meta.get("tentacle_id")
-    marker_written = _write_dispatched_subagent_marker(
-        tentacle_name=args.name,
-        scope=meta.get("scope", []),
-        dispatch_mode=args.output,
-        tentacle_id=tentacle_id,
-    )
-    if marker_written:
-        print(f"📌 Marker: {_DISPATCHED_MARKER_PATH}")
-        print(f"   Active until tentacle.py complete OR {_DISPATCHED_MARKER_TTL // 3600}h TTL.")
-        print("   Local enforcement surfaces (git hooks, preToolUse guards) may observe this.\n")
-
-    if args.output == "prompt":
-        # Output as a single dispatch prompt with all todos
-        print("─── DISPATCH PROMPT ───\n")
-        context_for_prompt = _render_dispatch_context(context, meta, bundle_dir)
-        prompt = _render_swarm_prompt(
-            args.name,
-            pending,
-            context_for_prompt,
-            specialist_profile_section=specialist_profile_section,
-            live_briefing_section=active_live_briefing_section,
-            dispatch_mode_section=dispatch_mode_section,
-            prompt_size_section=prompt_size_section,
-            bundle_section=bundle_section,
-            worktree_section=worktree_section,
-        )
-
-        print(prompt)
-
-        # Also output the task() call
-        print("\n─── COPILOT CLI DISPATCH ───\n")
-        escaped_prompt = prompt.replace('"', '\\"').replace("\n", "\\n")
-        print("task(")
-        print(f'    name="swarm-{args.name}",')
-        print(f'    agent_type="{agent_type}",')
-        print(f'    model="{model}",')
-        print('    mode="background",')
-        print(f'    description="Swarm: {args.name}",')
-        print('    prompt="""')
-        print(prompt)
-        print('"""')
-        print(")")
-
-    elif args.output == "parallel":
-        # Output one dispatch per todo (max parallelism)
-        print("─── PARALLEL DISPATCH (one agent per todo) ───\n")
-        print(f"Dispatch Mode: {dispatch_context_mode['summary']}")
-        if prompt_size.get("comparison_available"):
-            print(
-                "Prompt Size: "
-                f"{prompt_size['active_prompt_chars']} chars vs {prompt_size['full_context_prompt_chars']} chars "
-                f"({prompt_size['reduction_vs_full_context_percent']:.1f}% smaller than full-context fallback; "
-                f"target >= {POINTER_PROMPT_REDUCTION_TARGET_PERCENT:.1f}%)\n"
-            )
-        else:
-            print(f"Prompt Size: {prompt_size['active_prompt_chars']} chars (full-context inline fallback)\n")
-        for t in pending:
-            print(f"# Todo [{t['index']}]: {t['text']}")
-            print("task(")
-            print(f'    name="worker-{args.name}-{t["index"]}",')
-            print(f'    agent_type="{agent_type}",')
-            print(f'    model="{model}",')
-            print('    mode="background",')
-            print(f'    description="{t["text"][:50]}",')
-            print('    prompt="""')
-            print(f"## Tentacle: {args.name}")
-            print("")
-            print("### Context")
-            context_for_prompt = _render_dispatch_context(context, meta, bundle_dir)
-            print(f"{context_for_prompt[:900]}")
-            if specialist_profile_section:
-                print(specialist_profile_section.strip())
-            if active_live_briefing_section:
-                print(active_live_briefing_section.strip())
-            print(dispatch_mode_section.strip())
-            if bundle_section:
-                print(bundle_section.strip())
-            if worktree_section:
-                print(worktree_section.strip())
-            print("")
-            print("### Your Task")
-            print(f"{t['text']}")
-            print("")
-            print("### Guardrails")
-            print(
-                "- If a Bundle Path is present, read `manifest.json` first and use the bundle files as authoritative context"
-            )
-            print("- Stay within the scoped files only — DO NOT modify files outside your declared scope")
-            print("- **DO NOT run `git commit` or `git push`** — the orchestrator owns all git operations")
-            print("- **DO NOT widen your scope** without explicit escalation to the orchestrator")
-            print("- If the task requires files outside your scope, stop and write a scope escalation note to handoff")
-            print("")
-            print("### Cross-review (required before handoff)")
-            print("Re-read every file you modified and confirm correctness before writing handoff.")
-            print("")
-            print("### When done")
-            print(f'python3 ~/.copilot/tools/tentacle.py todo "{args.name}" done {t["index"]}')
-            print(
-                f'python3 ~/.copilot/tools/tentacle.py handoff "{args.name}" "Completed: {t["text"]}" --status DONE --changed-file <path1> --changed-file <path2> --learn'
-            )
-            print(
-                "# Status: DONE | BLOCKED | TOO_BIG | AMBIGUOUS | REGRESSED  (use --changed-file once per modified file)"
-            )
-            print('"""')
-            print(")\n")
-
-    elif args.output == "json":
-        # Output structured JSON for programmatic use
-        dispatch = {
-            "tentacle": args.name,
-            "agent_type": agent_type,
-            "model": model,
-            "context_file": str(context_path),
-            "pending_todos": [{"index": t["index"], "text": t["text"]} for t in pending],
-            "execution_guidance": {
-                "git_ops": "Do not run git commit or git push — the orchestrator owns all git operations",
-                "scope": "Stay within declared files — do not widen scope without escalating to the orchestrator",
-                "escalation": "If scope is insufficient, stop and write a scope escalation note to handoff",
-                "context_bundle": (
-                    "Pointer-based bundles are the default. Read bundle_path/manifest.json first, then "
-                    "context-packet.md, session-metadata.md, and recall-pack.json before editing."
-                    if bundle_dir is not None
-                    else "Full-context inline fallback is active; rely on context_file and the inline prompt."
-                ),
-            },
-            "dispatch_context_mode": dispatch_context_mode,
-            "prompt_size": prompt_size,
-            "marker_state": _get_marker_state(),
-        }
-        if agent_profile:
-            dispatch["agent_profile"] = _agent_profile_meta(agent_profile)
-        if bundle_dir is not None:
-            dispatch["bundle_path"] = str(bundle_dir)
-            dispatch["context_packet_path"] = str(bundle_dir / "context-packet.md")
-        if wt_path_str is not None:
-            dispatch["worktree_path"] = wt_path_str
-        print(json.dumps(dispatch, indent=2))
+# Review dispatch command lives in _tentacle_review.py.
 
 
-def cmd_dispatch_reviewer(args) -> None:
-    """Generate a fresh-context reviewer prompt and minimal reviewer bundle."""
-    tentacles = get_tentacles_dir(args.session_dir)
-    tentacle_dir = _validate_tentacle_name(args.name, tentacles)
-    if not tentacle_dir.exists():
-        print(f"ERROR: Tentacle '{args.name}' not found.", file=sys.stderr)
-        sys.exit(1)
-
-    _require_done_handoff(tentacle_dir, "dispatch-reviewer")
-
-    meta_path = tentacle_dir / "meta.json"
-    meta = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    try:
-        bundle_info = _build_reviewer_bundle(tentacle_dir, args.name, meta)
-    except RuntimeError as exc:
-        print(f"ERROR: {exc}", file=sys.stderr)
-        sys.exit(1)
-
-    reviewer_meta = {
-        **(meta.get("reviewer") or {}),
-        **bundle_info,
-        "updated_at": datetime.now(timezone.utc).isoformat(),
-    }
-    meta["reviewer"] = reviewer_meta
-    meta_path.write_text(json.dumps(meta, indent=2) + "\n", encoding="utf-8")
-
-    agent_type = getattr(args, "agent_type", None) or "code-review"
-    model = getattr(args, "model", None) or "claude-sonnet-4.6"
-    prompt = _render_dispatch_reviewer_prompt(args.name, bundle_info)
-
-    if getattr(args, "output", "prompt") == "json":
-        print(
-            json.dumps(
-                {
-                    "tentacle": args.name,
-                    "agent_type": agent_type,
-                    "model": model,
-                    "bundle_path": bundle_info["bundle_path"],
-                    "manifest_path": bundle_info["manifest_path"],
-                    "review_context_path": bundle_info["review_context_path"],
-                    "diff_path": bundle_info["diff_path"],
-                    "spec_path": bundle_info["spec_path"],
-                    "findings_path": bundle_info["findings_path"],
-                    "diff_source": bundle_info["diff_source"],
-                    "result_contract": {
-                        "safe_to_merge": "SAFE_TO_MERGE: YES|NO",
-                        "blockers_heading": "### BLOCKERS",
-                        "warnings_heading": "### WARNINGS",
-                    },
-                    "review_loop_integration": {
-                        "safe_result": "lane may continue",
-                        "blocker_result": "review-loop converts blockers into actionable follow-up context",
-                        "warnings": "visible but non-blocking",
-                    },
-                    "prompt": prompt,
-                },
-                indent=2,
-            )
-        )
-        return
-
-    print(f"🧾 Fresh-context reviewer bundle: {bundle_info['bundle_path']}")
-    print(f"📝 Findings file: {bundle_info['findings_path']}")
-    print(f"🌿 Diff source: {bundle_info['diff_source']}")
-    print("")
-    print("─── REVIEWER PROMPT ───\n")
-    print(prompt)
-    print("\n─── COPILOT CLI DISPATCH ───\n")
-    print("task(")
-    print(f'    name="reviewer-{args.name}",')
-    print(f'    agent_type="{agent_type}",')
-    print(f'    model="{model}",')
-    print('    mode="background",')
-    print(f'    description="Review: {args.name}",')
-    print('    prompt="""')
-    print(prompt)
-    print('"""')
-    print(")")
-
-
-def cmd_next_step(args):
-    """Show the grounded next step for a tentacle: first pending todo + checkpoint/briefing context.
-
-    Read-only — does not mutate tentacle state.
-    """
-    tentacles = get_tentacles_dir(args.session_dir)
-    tentacle_dir = _validate_tentacle_name(args.name, tentacles)
-
-    if not tentacle_dir.exists():
-        print(f"ERROR: Tentacle '{args.name}' not found.", file=sys.stderr)
-        sys.exit(1)
-
-    meta_path = tentacle_dir / "meta.json"
-    todo_path = tentacle_dir / "todo.md"
-
-    meta = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-    todos = parse_todos(todo_path.read_text(encoding="utf-8")) if todo_path.exists() else []
-    pending = [t for t in todos if not t["done"]]
-    done_count = sum(1 for t in todos if t["done"])
-
-    # Load checkpoint context unless suppressed
-    checkpoint_text = ""
-    if not getattr(args, "no_checkpoint", False):
-        checkpoint_text = _load_latest_checkpoint_context()
-
-    # Load briefing only when explicitly requested
-    briefing_text = ""
-    if getattr(args, "briefing", False):
-        fallback = meta.get("description", "") or args.name.replace("-", " ")
-        briefing_text = _run_briefing_for_task(args.name, fallback_query=fallback)
-
-    fmt = getattr(args, "format", "text")
-    if fmt == "json":
-        output = {
-            "tentacle": args.name,
-            "status": meta.get("status", "idle"),
-            "todos_done": done_count,
-            "todos_total": len(todos),
-            "pending": [{"index": t["index"], "text": t["text"]} for t in pending],
-            "next_step": pending[0]["text"] if pending else None,
-            "checkpoint_context": checkpoint_text or None,
-            "briefing": briefing_text or None,
-        }
-        print(json.dumps(output, indent=2))
-        return
-
-    # Human-readable output
-    print(f"🎯 Next step for '{args.name}'")
-    print(f"   Status: {meta.get('status', 'idle')} | Progress: {done_count}/{len(todos)} done")
-    print()
-
-    if not pending:
-        print("✅ All todos done! Nothing pending.")
-        if checkpoint_text:
-            print()
-            print(checkpoint_text)
-        return
-
-    next_todo = pending[0]
-    print(f"▶  [{next_todo['index']}] {next_todo['text']}")
-
-    if getattr(args, "all", False) and len(pending) > 1:
-        print(f"\n   Also pending ({len(pending) - 1} more):")
-        for t in pending[1:]:
-            print(f"   ☐ [{t['index']}] {t['text']}")
-
-    if checkpoint_text:
-        print()
-        print(checkpoint_text)
-
-    if briefing_text:
-        print()
-        print("### Knowledge Briefing")
-        print(briefing_text)
+# Next-step command lives in _tentacle_dispatch.py.
 
 
 def cmd_delete(args):
@@ -5701,119 +3121,7 @@ def cmd_delete(args):
     print(f"🗑️  Tentacle '{args.name}' deleted.")
 
 
-def cmd_bundle(args):
-    """Materialize a per-run context bundle for a tentacle subagent."""
-    tentacles = get_tentacles_dir(args.session_dir)
-    tentacle_dir = _validate_tentacle_name(args.name, tentacles)
-    json_output = getattr(args, "output", "text") == "json"
-
-    if not tentacle_dir.exists():
-        print(f"ERROR: Tentacle '{args.name}' not found.", file=sys.stderr)
-        sys.exit(1)
-
-    meta_path = tentacle_dir / "meta.json"
-    meta = json.loads(meta_path.read_text(encoding="utf-8")) if meta_path.exists() else {}
-
-    # Fetch briefing + recall pack
-    fallback = meta.get("description", "") or args.name.replace("-", " ")
-    recall_pack_data, recall_source_mode = _fetch_recall_pack_json(args.name, fallback_query=fallback)
-    briefing_text = ""
-    if not getattr(args, "no_briefing", False):
-        if not json_output:
-            print(f"🧠 Fetching briefing for '{args.name}'...")
-        briefing_text = _render_recall_payload(
-            args.name,
-            recall_pack_data,
-            recall_source_mode,
-        )
-        if not json_output:
-            if briefing_text:
-                print(f"   ✅ Briefing: {len(briefing_text)} chars")
-            else:
-                print("   ℹ️  No briefing data — placeholder will be written")
-    if recall_pack_data and not json_output:
-        print(f"   ✅ Recall pack: {recall_source_mode} ({len(json.dumps(recall_pack_data))} chars)")
-
-    # Load checkpoint
-    checkpoint_text = ""
-    if not getattr(args, "no_checkpoint", False):
-        checkpoint_text = _load_latest_checkpoint_context()
-
-    # Worktree preparation (--worktree flag)
-    wt_path_str: str | None = None
-    if getattr(args, "worktree", False):
-        if not json_output:
-            print(f"🌿 Preparing worktree for '{args.name}'...")
-        git_root = find_git_root()
-        wt_state = _worktree_prepare(tentacle_dir, args.name, git_root)
-        if wt_state["prepared"]:
-            wt_path_str = wt_state["path"]
-            action = "reused" if wt_state.get("reused") else "prepared"
-            if not json_output:
-                print(f"   ✅ Worktree {action}: {wt_path_str}")
-        else:
-            if not json_output:
-                print(f"   ⚠️  Worktree prepare failed: {wt_state.get('error', 'unknown')}")
-
-    # Gather goal context text if this tentacle is linked to a goal
-    goal_context_text = ""
-    context_packet_goal_context_text = ""
-    bundle_prior_handoffs: list[dict] = []
-    try:
-        goal_state = _goal_load(tentacles)
-        if goal_state and args.name in (goal_state.get("tentacles") or []):
-            goal_context_text = _goal_render_continuation_context(goal_state, tentacles)
-            context_packet_goal_context_text = _goal_render_continuation_context(
-                goal_state,
-                tentacles,
-                include_prior_handoffs=False,
-            )
-            bundle_prior_handoffs = _goal_collect_prior_handoffs(goal_state, tentacles)
-            if not json_output:
-                print(f"   ✅ Goal context: {len(goal_context_text)} chars")
-    except Exception:
-        pass
-
-    bundle_dir = _build_runtime_bundle(
-        tentacle_dir=tentacle_dir,
-        name=args.name,
-        briefing_text=briefing_text,
-        checkpoint_text=checkpoint_text,
-        worktree_path=wt_path_str,
-        recall_pack_data=recall_pack_data,
-        recall_source_mode=recall_source_mode,
-        goal_context_text=goal_context_text,
-        context_packet_goal_context_text=context_packet_goal_context_text,
-        prior_handoffs=bundle_prior_handoffs,
-    )
-
-    # Write dispatched-subagent-active marker when materializing a bundle
-    tentacle_id = meta.get("tentacle_id")
-    _write_dispatched_subagent_marker(
-        tentacle_name=args.name,
-        scope=meta.get("scope", []),
-        dispatch_mode="bundle",
-        tentacle_id=tentacle_id,
-    )
-
-    if json_output:
-        manifest_path = bundle_dir / "manifest.json"
-        manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
-        out = {
-            "bundle_path": str(bundle_dir),
-            "marker_state": _get_marker_state(),
-            **manifest,
-        }
-        if wt_path_str:
-            out["worktree_path"] = wt_path_str
-        print(json.dumps(out, indent=2))
-    else:
-        print(f"📦 Bundle materialized: {bundle_dir}")
-        if wt_path_str:
-            print(f"🌿 Worktree: {wt_path_str}")
-        print(f"📌 Marker: {_DISPATCHED_MARKER_PATH}")
-        for f in sorted(bundle_dir.iterdir()):
-            print(f"   {f.name} ({f.stat().st_size} bytes)")
+# Bundle command lives in _tentacle_dispatch.py.
 
 
 # ---------------------------------------------------------------------------
@@ -6076,6 +3384,73 @@ def _pr_runtime_subprocess_safe(
     return _pr_run_subprocess_safe(cmd, **kwargs)
 
 
+def _dispatch_runtime_fetch_recall_pack_json(task_id: str, fallback_query: str = "") -> tuple[dict, str | None]:
+    return _fetch_recall_pack_json(task_id, fallback_query=fallback_query)
+
+
+def _dispatch_runtime_run_briefing_for_task(task_id: str, fallback_query: str = "") -> str:
+    return _run_briefing_for_task(task_id, fallback_query=fallback_query)
+
+
+def _dispatch_runtime_build_runtime_bundle(*args, **kwargs) -> Path:
+    return _build_runtime_bundle(*args, **kwargs)
+
+
+def _review_runtime_build_reviewer_bundle(tentacle_dir: Path, name: str, meta: dict) -> dict:
+    return _build_reviewer_bundle(tentacle_dir, name, meta)
+
+
+def _configure_dispatch_runtime() -> None:
+    _dispatch.configure_dispatch_runtime(
+        _runtime_BRIEFING_PY=lambda: BRIEFING_PY,
+        _runtime_CHECKPOINT_RESTORE_PY=lambda: CHECKPOINT_RESTORE_PY,
+        _runtime_HANDOFF_TRIAGE_STATUSES=HANDOFF_TRIAGE_STATUSES,
+        _runtime_find_git_root=lambda: find_git_root(),
+        _runtime_get_tentacles_dir=lambda session_dir=None: get_tentacles_dir(session_dir),
+        _runtime_parse_todos=lambda text: parse_todos(text),
+        _runtime_validate_tentacle_name=lambda name, tentacles: _validate_tentacle_name(name, tentacles),
+        _runtime_worktree_prepare=lambda tentacle_dir, name, git_root=None: _worktree_prepare(
+            tentacle_dir, name, git_root
+        ),
+        _runtime_write_dispatched_subagent_marker=lambda **kwargs: _write_dispatched_subagent_marker(**kwargs),
+        _runtime_get_marker_state=lambda: _get_marker_state(),
+        _runtime_goal_load=lambda tentacles: _goal_load(tentacles),
+        _runtime_goal_render_continuation_context=lambda goal_state, tentacles, include_prior_handoffs=True: (
+            _goal_render_continuation_context(goal_state, tentacles, include_prior_handoffs=include_prior_handoffs)
+        ),
+        _runtime_goal_collect_prior_handoffs=lambda goal_state, tentacles: _goal_collect_prior_handoffs(
+            goal_state, tentacles
+        ),
+        _runtime_fetch_recall_pack_json=_dispatch_runtime_fetch_recall_pack_json,
+        _runtime_run_briefing_for_task=_dispatch_runtime_run_briefing_for_task,
+        _runtime_load_latest_checkpoint_context=lambda: _load_latest_checkpoint_context(),
+        _runtime_build_runtime_bundle=_dispatch_runtime_build_runtime_bundle,
+    )
+
+
+def _configure_review_runtime() -> None:
+    _review.configure_review_runtime(
+        _runtime_find_git_root=lambda: find_git_root(),
+        _runtime_get_tentacles_dir=lambda session_dir=None: get_tentacles_dir(session_dir),
+        _runtime_validate_tentacle_name=lambda name, tentacles: _validate_tentacle_name(name, tentacles),
+        _runtime_scope_items=lambda meta: _scope_items(meta),
+        _runtime_parse_bullet_list=lambda text: _parse_bullet_list(text),
+        _runtime_parse_handoff_status=lambda text: _parse_handoff_status(text),
+        _runtime_parse_handoff_changed_files=lambda text: _parse_handoff_changed_files(text),
+        _runtime_parse_rich_handoff_sections=lambda text: _parse_rich_handoff_sections(text),
+        _runtime_tentacle_slug=lambda name: _tentacle_slug(name),
+        _runtime_update_meta_worktree=lambda tentacle_dir, worktree_state: _update_meta_worktree(
+            tentacle_dir, worktree_state
+        ),
+        _runtime_goal_load=lambda tentacles: _goal_load(tentacles),
+        _runtime_cmd_goal_link=lambda args, tentacles: _cmd_goal_link(args, tentacles),
+        _runtime_cmd_swarm=lambda args: cmd_swarm(args),
+        _runtime_run_and_record_verification=lambda **kwargs: _run_and_record_verification(**kwargs),
+        _runtime_reviewer_collect_diff=lambda meta: _reviewer_collect_diff(meta),
+        _runtime_build_reviewer_bundle=_review_runtime_build_reviewer_bundle,
+    )
+
+
 def _configure_pr_runtime() -> None:
     _pr.configure_pr_runtime(
         GOAL_STATUS_COMPLETED=GOAL_STATUS_COMPLETED,
@@ -6112,6 +3487,8 @@ def _configure_goal_runtime() -> None:
 
 _configure_goal_runtime()
 _configure_pr_runtime()
+_configure_dispatch_runtime()
+_configure_review_runtime()
 
 # ---------------------------------------------------------------------------
 # SEAM: core-cli argparse boundary
