+++ tests/test_quality_gates.py
@@ -849,6 +849,8 @@ CI_RUFF_FILES = [
     "_tentacle_core.py",
     "_tentacle_goal.py",
     "_tentacle_pr.py",
+    "_tentacle_dispatch.py",
+    "_tentacle_review.py",
     "checkpoint-diff.py",
     "checkpoint-restore.py",
     "checkpoint-save.py",
@@ -1370,6 +1372,8 @@ def test_architecture_md_ruff_surface():
         "_tentacle_core.py",
         "_tentacle_goal.py",
         "_tentacle_pr.py",
+        "_tentacle_dispatch.py",
+        "_tentacle_review.py",
         "tests/test_browse_search_v2.py",
         "browse/",
         "hooks/",
